Abstract Annotation Group

This is an abstract class to augment the base TiledbGroup class with fields necessary for the child Annotation- classes. Currently it only adds the dimension_name field.

Super classes

tiledbsoma::TileDBObject -> tiledbsoma::TileDBGroup -> AnnotationGroup

Public fields

dimension_name

Optional name of the dimension shared by all arrays within the group (typically obs_id or var_id).

Methods

Public methods

AnnotationGroup$new()

AnnotationGroup$set_query()

AnnotationGroup$reset_query()

AnnotationGroup$clone()

Inherited methods

tiledbsoma::TileDBObject$class()

tiledbsoma::TileDBObject$exists()

tiledbsoma::TileDBGroup$add_member()

tiledbsoma::TileDBGroup$add_metadata()

tiledbsoma::TileDBGroup$count_members()

tiledbsoma::TileDBGroup$get_member()

tiledbsoma::TileDBGroup$get_members()

tiledbsoma::TileDBGroup$get_metadata()

tiledbsoma::TileDBGroup$group_exists()

tiledbsoma::TileDBGroup$list_member_uris()

tiledbsoma::TileDBGroup$list_members()

tiledbsoma::TileDBGroup$list_object_uris()

tiledbsoma::TileDBGroup$list_objects()

tiledbsoma::TileDBGroup$print()

tiledbsoma::TileDBGroup$remove_member()

tiledbsoma::TileDBGroup$tiledb_group()

Method new()

Create a new TileDBGroup-based Annotation class.

Usage

AnnotationGroup$new(
  uri,
  dimension_name = NULL,
  verbose = TRUE,
  config = NULL,
  ctx = NULL
)

Arguments

uri

URI for the TileDB group.

dimension_name

Optional name of the dimension shared by all of the group's member arrays.

verbose

Print status messages

config

optional configuration

ctx

optional tiledb context

Method set_query()

Set dimension values to slice from the array members.

Usage

AnnotationGroup$set_query(dims = NULL, attr_filter = NULL)

Arguments

dims

a named list of character vectors. Each must correspond to a dimension shared by all array members.

attr_filter

a TileDB query condition for attribute filtering pushdown.

Method reset_query()

Reset the group member queries.

Usage

AnnotationGroup$reset_query(dims = TRUE, attr_filter = TRUE)

Arguments

dims

Clear the defined dimension ranges?

attr_filter

Clear the defined attribute filters?

Returns

NULL

Method clone()

The objects of this class are cloneable with this method.

Usage

AnnotationGroup$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.