Single-cell Annotation Matrix Group

Class for representing a TileDB group containing one or more AnnotationPairwiseMatrix arrays that share a common dimension name.

Super classes

tiledbsoma::TileDBObject -> tiledbsoma::TileDBGroup -> tiledbsoma::AnnotationGroup -> AnnotationPairwiseMatrixGroup

Methods

Public methods

AnnotationPairwiseMatrixGroup$add_matrix()

AnnotationPairwiseMatrixGroup$add_seurat_graph()

AnnotationPairwiseMatrixGroup$clone()

Inherited methods

tiledbsoma::TileDBObject$class()

tiledbsoma::TileDBObject$exists()

tiledbsoma::TileDBGroup$add_member()

tiledbsoma::TileDBGroup$add_metadata()

tiledbsoma::TileDBGroup$count_members()

tiledbsoma::TileDBGroup$get_member()

tiledbsoma::TileDBGroup$get_members()

tiledbsoma::TileDBGroup$get_metadata()

tiledbsoma::TileDBGroup$group_exists()

tiledbsoma::TileDBGroup$list_member_uris()

tiledbsoma::TileDBGroup$list_members()

tiledbsoma::TileDBGroup$list_object_uris()

tiledbsoma::TileDBGroup$list_objects()

tiledbsoma::TileDBGroup$print()

tiledbsoma::TileDBGroup$remove_member()

tiledbsoma::TileDBGroup$tiledb_group()

tiledbsoma::AnnotationGroup$initialize()

tiledbsoma::AnnotationGroup$reset_query()

tiledbsoma::AnnotationGroup$set_query()

Method add_matrix()

Add a new AnnotationPairwiseMatrix array to the group.

Usage

AnnotationPairwiseMatrixGroup$add_matrix(data, name, metadata = NULL)

Arguments

data

a matrix of annotation data to ingest. The matrix rows must be aligned to the dimension indicated by the group's dimension_name.

name

Name of the new pairwise annotation matrix.

metadata

Named list of metadata to add.

Method add_seurat_graph()

Convert a SeuratObject::Graph object to AnnotationPairwiseMatrix.

Usage

AnnotationPairwiseMatrixGroup$add_seurat_graph(object, technique)

Arguments

object

A SeuratObject::Graph object.

technique

Name of the technique used to generate the graph (typically, nn or snn).

Details

On-Disk Format

Arrays are named graph_<technique>.

Metadata

assay_used: Name of the assay used to generate the graph.

graph_technique: Name of the technique used to generate the graph. used.

Method clone()

The objects of this class are cloneable with this method.

Usage

AnnotationPairwiseMatrixGroup$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.