Single-cell Assay Matrix

Base class for 2D sparse matrix-like data with string dimensions. An AssayMatrix may contain one or more "layers" (i.e., additional measurements that share the same dimensions and non-empty coordinates.

Used for the X field of SOMA.

Super classes

tiledbsoma::TileDBObject -> tiledbsoma::TileDBArray -> tiledbsoma::AnnotationArray -> AssayMatrix

Public fields

verbose

Print status messages

Methods

Public methods

AssayMatrix$from_matrix()

AssayMatrix$from_dataframe()

AssayMatrix$to_dataframe()

AssayMatrix$to_matrix()

AssayMatrix$clone()

Inherited methods

tiledbsoma::TileDBObject$class()

tiledbsoma::TileDBObject$exists()

tiledbsoma::TileDBArray$add_metadata()

tiledbsoma::TileDBArray$array_exists()

tiledbsoma::TileDBArray$attributes()

tiledbsoma::TileDBArray$attrnames()

tiledbsoma::TileDBArray$dimensions()

tiledbsoma::TileDBArray$dimnames()

tiledbsoma::TileDBArray$fragment_count()

tiledbsoma::TileDBArray$get_metadata()

tiledbsoma::TileDBArray$initialize()

tiledbsoma::TileDBArray$print()

tiledbsoma::TileDBArray$reset_query()

tiledbsoma::TileDBArray$schema()

tiledbsoma::TileDBArray$set_query()

tiledbsoma::TileDBArray$tiledb_array()

Method from_matrix()

Ingest assay data from a sparse matrix

Usage

AssayMatrix$from_matrix(x, index_cols, value_col = "value", transpose = FALSE)

Arguments

x

any matrix-like object coercible to a TsparseMatrix with string dimensions.

index_cols

Names to use for the TileDB array's dimensions that will contain the matrix row/column names.

value_col

Name to use for the TileDB array's attribute that will contain the matrix values.

transpose

If TRUE, the order of the new TileDB array's dimensions are reversed relative to the names defined by index_cols.

Transposing Dimensions

The transpose argument allows you to ingest a matrix into a TileDB array with the dimension order reversed relative to the original matrix shape. For example, consider matrix mat with dimensions i and j:

mat <- matrix(1:6, nrow = 3, dimnames = list(LETTERS[1:3], letters[1:2]))
mat
#>   a b
#> A 1 4
#> B 2 5
#> C 3 6

By default AssayMatrix will create a TileDB array will have dimensions i/j (i.e., the same order as the matrix):

amat1 <- AssayMatrix$new(uri = "mem://assaymatrix1", verbose = FALSE)
amat1$from_matrix(mat, index_cols = c("i", "j"))
amat1$to_dataframe()
#>   i j value
#> 1 A a     1
#> 2 A b     4
#> 3 B a     2
#> 4 B b     5
#> 5 C a     3
#> 6 C b     6

If you set transpose = TRUE, the TileDB array will have dimensions j/i instead:

amat2 <- AssayMatrix$new(uri = "mem://assaymatrix2", verbose = FALSE)
amat2$from_matrix(mat, index_cols = c("i", "j"), transpose = TRUE)
amat2$to_dataframe()
#>   j i value
#> 1 a A     1
#> 2 a B     2
#> 3 a C     3
#> 4 b A     4
#> 5 b B     5
#> 6 b C     6

Method from_dataframe()

Ingest assay data from a COO-formatted data frame

Usage

AssayMatrix$from_dataframe(x, index_cols)

Arguments

x

a data.frame

index_cols

A column index, either numeric with a column index, or character with a column name, identifying the 2 index columns. All other columns are ingested as attributes.

Method to_dataframe()

Retrieve the assay data from TileDB

Usage

AssayMatrix$to_dataframe(attrs = NULL, batch_mode = FALSE)

Arguments

attrs

Specify one or more attributes to retrieve. If NULL, all attributes are retrieved.

batch_mode

logical, if TRUE, batch query mode is enabled, which provides the ability to detect partial query results and resubmit until all results are retrieved.

Returns

A Matrix::dgTMatrix.

Method to_matrix()

Retrieve assay data from TileDB as a 2D sparse matrix.

Usage

AssayMatrix$to_matrix(attr = NULL, batch_mode = FALSE, transpose = FALSE)

Arguments

attr

The name of the attribute layer to retrieve. If NULL, the first layer is returned.

batch_mode

logical, if TRUE, batch query mode is enabled, which provides the ability to detect partial query results and resubmit until all results are retrieved.

transpose

If TRUE, the order of the matrix's dimensions are reversed relative to the TileDB array's dimension name. dimensions are reversed relative to the names defined by index_cols.

Returns

A Matrix::dgTMatrix.

Method clone()

The objects of this class are cloneable with this method.

Usage

AssayMatrix$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.