Single-cell Assay Matrix Group

Class for representing a TileDB group containing one or more AssayMatrix arrays that share the same dimensions.

Super classes

tiledbsoma::TileDBObject -> tiledbsoma::TileDBGroup -> tiledbsoma::AnnotationGroup -> AssayMatrixGroup

Methods

Public methods

AssayMatrixGroup$add_assay_matrix()

AssayMatrixGroup$clone()

Inherited methods

tiledbsoma::TileDBObject$class()

tiledbsoma::TileDBObject$exists()

tiledbsoma::TileDBGroup$add_member()

tiledbsoma::TileDBGroup$add_metadata()

tiledbsoma::TileDBGroup$count_members()

tiledbsoma::TileDBGroup$get_member()

tiledbsoma::TileDBGroup$get_members()

tiledbsoma::TileDBGroup$get_metadata()

tiledbsoma::TileDBGroup$group_exists()

tiledbsoma::TileDBGroup$list_member_uris()

tiledbsoma::TileDBGroup$list_members()

tiledbsoma::TileDBGroup$list_object_uris()

tiledbsoma::TileDBGroup$list_objects()

tiledbsoma::TileDBGroup$print()

tiledbsoma::TileDBGroup$remove_member()

tiledbsoma::TileDBGroup$tiledb_group()

tiledbsoma::AnnotationGroup$initialize()

tiledbsoma::AnnotationGroup$reset_query()

tiledbsoma::AnnotationGroup$set_query()

Method add_assay_matrix()

Add a new AssayMatrix array to the group.

Usage

AssayMatrixGroup$add_assay_matrix(
  data,
  name,
  value_col = "value",
  metadata = NULL,
  transpose = FALSE
)

Arguments

data

a matrix of annotation data to ingest. The matrix rows must be aligned to the dimension indicated by the group's dimension_name.

name

Name for the new array, nested with the group's URI.

value_col

Name to use for the TileDB array's attribute that will contain the matrix values.

metadata

Named list of metadata to add.

transpose

If TRUE, the order of the new TileDB array's dimensions are reversed relative to the names defined by AssayMatrixGroup's dimension_name.

Method clone()

The objects of this class are cloneable with this method.

Usage

AssayMatrixGroup$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.