SOMA Blockwise Read Iterator for Sparse Matrices

Class that allows for blockwise read iteration of SOMA reads as sparse matrices

Super classes

tiledbsoma::ReadIter -> tiledbsoma::BlockwiseReadIterBase -> BlockwiseSparseReadIter

Active bindings

repr

Representation of the sparse matrix to return

Methods

Public methods

BlockwiseSparseReadIter$new()

BlockwiseSparseReadIter$concat()

BlockwiseSparseReadIter$clone()

Inherited methods

tiledbsoma::BlockwiseReadIterBase$read_complete()

tiledbsoma::BlockwiseReadIterBase$read_next()

Method new()

Create

Usage

BlockwiseSparseReadIter$new(
  sr,
  array,
  coords,
  axis,
  ...,
  repr = "T",
  reindex_disable_on_axis = NA
)

Arguments

sr

SOMA read pointer

array

Underlying SOMASparseNDArray

coords

Named list of CoordsStrider objects; must be named after array$dimnames()

axis

Axis to iterate over in a blockwise manner

...

Ignored

repr

Representation of the sparse matrix to return; choose from:

“T”: returns a TsparseMatrix

“R”: returns an RsparseMatrix

“C”: returns a CsparseMatrix

Note: passing repr of “R” or “C” are only available if re-indexing is enabled on axes 0 or 1, respectively

reindex_disable_on_axis

Additional axes that will not be re-indexed; the following values may be used as shorthands for common settings:

“TRUE”: disable re-indexing on all axes

“NA”: re-index only on axis, disable re-indexing on all others

“FALSE”: re-index on all axes, do not disable re-indexing

Method concat()

Concatenate the remainder of the blockwise iterator

Usage

BlockwiseSparseReadIter$concat()

Returns

A sparse matrix (determined by self$repr) with the remainder of the iterator

Method clone()

The objects of this class are cloneable with this method.

Usage

BlockwiseSparseReadIter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.