Ephemeral Collection Base

Base class for ephemeral collections; ephemeral collections are equivalent to SOMA collections but are stored in-memory instead of on-disk

Super classes

tiledbsoma::TileDBObject -> tiledbsoma::TileDBGroup -> tiledbsoma::SOMACollectionBase -> EphemeralCollectionBase

Active bindings

uri

“ephemeral-collection:<MEMORY_ADDRESS>”

soma_type

Dummy field for ephemeral objects for compatibility with SOMA collections

platform_config

Dummy field for ephemeral objects for compatibility with SOMA collections

tiledbsoma_ctx

Dummy field for ephemeral objects for compatibility with SOMA collections

object

Dummy field for ephemeral objects for compatibility with SOMA collections

Methods

Public methods

EphemeralCollectionBase$new()

EphemeralCollectionBase$create()

EphemeralCollectionBase$open()

EphemeralCollectionBase$close()

EphemeralCollectionBase$exists()

EphemeralCollectionBase$print()

EphemeralCollectionBase$get_tiledb_config()

EphemeralCollectionBase$length()

EphemeralCollectionBase$names()

EphemeralCollectionBase$set()

EphemeralCollectionBase$get()

EphemeralCollectionBase$remove()

EphemeralCollectionBase$set_metadata()

EphemeralCollectionBase$get_metadata()

EphemeralCollectionBase$add_new_collection()

EphemeralCollectionBase$add_new_dataframe()

EphemeralCollectionBase$add_new_dense_ndarray()

EphemeralCollectionBase$add_new_sparse_ndarray()

EphemeralCollectionBase$clone()

Inherited methods

tiledbsoma::TileDBObject$class()

tiledbsoma::TileDBObject$is_open()

tiledbsoma::TileDBObject$mode()

tiledbsoma::TileDBObject$reopen()

tiledbsoma::TileDBGroup$to_data_frame()

tiledbsoma::TileDBGroup$to_list()

Method new()

Create an ephemeral collection

Usage

EphemeralCollectionBase$new(...)

Arguments

...

Ignored

Method create()

Create a new, empty ephemeral collection

Usage

EphemeralCollectionBase$create()

Returns

Returns a new ephemeral collection of class class(self)

Method open()

Dummy method for ephemeral cobjects for compatibility with SOMA collections

Usage

EphemeralCollectionBase$open(mode)

Arguments

mode

Ignored for ephemeral objects

Returns

Throws an error as this method is not supported by ephemeral objects

Method close()

Dummy method for ephemeral cobjects for compatibility with SOMA collections

Usage

EphemeralCollectionBase$close()

Returns

Invisibly returns NULL

Method exists()

Dummy method for ephemeral cobjects for compatibility with SOMA collections

Usage

EphemeralCollectionBase$exists()

Returns

Returns FALSE as ephemeral collections do not exist on-disk

Method print()

Special method for printing object representation to console

Usage

EphemeralCollectionBase$print()

Returns

Prints details about the ephemeral collection and invisibly returns itself

Method get_tiledb_config()

Dummy method for ephemeral cobjects for compatibility with SOMA collections

Usage

EphemeralCollectionBase$get_tiledb_config(param = NULL)

Arguments

param

Ignored for ephemeral objects

Returns

Returns NULL as ephemeral collections do not have an on-disk configuration

Method length()

Retrieve the number of items in the collection

Usage

EphemeralCollectionBase$length()

Returns

The length of the collection

Method names()

Retrieve the names of members. (lifecycle: experimental)

Usage

EphemeralCollectionBase$names()

Returns

A character vector of member names.

Method set()

Add object to an ephemeral collection

Usage

EphemeralCollectionBase$set(object, name = NULL, relative = NULL)

Arguments

object

A TileDB object (eg. TileDBGroup) to add to the collection

name

A name to add object as

relative

Ignored for ephemeral objects

Returns

[chainable] Invisibly returns self with object added as name

Method get()

Get objects from an ephemeral collection

Usage

EphemeralCollectionBase$get(name)

Arguments

name

Name of object in the collection to get

Returns

The object named name

Method remove()

Remove objects from an ephemeral collection

Usage

EphemeralCollectionBase$remove(name)

Arguments

name

Name of object to remove from the collection

Returns

[chainable] Invisibly returns self with the object at name removed

Method set_metadata()

Dummy method for ephemeral cobjects for compatibility with SOMA collections

Usage

EphemeralCollectionBase$set_metadata(metadata)

Arguments

metadata

Ignored for ephemeral objects

Returns

Throws an error as this method is not supported by ephemeral objects

Method get_metadata()

Dummy method for ephemeral cobjects for compatibility with SOMA collections

Usage

EphemeralCollectionBase$get_metadata(key = NULL)

Arguments

key

Ignored for ephemeral objects

Returns

An empty list

Method add_new_collection()

Dummy method for ephemeral cobjects for compatibility with SOMA collections

Usage

EphemeralCollectionBase$add_new_collection(object, key)

Arguments

object, key

Ignored for ephemeral objects

Returns

Throws an error as this method is not supported by ephemeral objects

Method add_new_dataframe()

Dummy method for ephemeral cobjects for compatibility with SOMA collections

Usage

EphemeralCollectionBase$add_new_dataframe(key, schema, index_column_names)

Arguments

key, schema, index_column_names

Ignored for ephemeral objects

Returns

Throws an error as this method is not supported by ephemeral objects

Method add_new_dense_ndarray()

Dummy method for ephemeral cobjects for compatibility with SOMA collections

Usage

EphemeralCollectionBase$add_new_dense_ndarray(key, type, shape)

Arguments

key, type, shape

Ignored for ephemeral objects

Returns

Throws an error as this method is not supported by ephemeral objects

Method add_new_sparse_ndarray()

Dummy method for ephemeral cobjects for compatibility with SOMA collections

Usage

EphemeralCollectionBase$add_new_sparse_ndarray(key, type, shape)

Arguments

key, type, shape

Ignored for ephemeral objects

Returns

Throws an error as this method is not supported by ephemeral objects

Method clone()

The objects of this class are cloneable with this method.

Usage

EphemeralCollectionBase$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.