SOMAExperiment Axis Query Result Indexer

Index obs/var soma_joinids for a given query result.

Retrieve the index of the given obs or var coordinates in the query result. Coordinates outside of the query result will return arrow::null().

Methods

Public methods

SOMAAxisIndexer$new()

SOMAAxisIndexer$by_obs()

SOMAAxisIndexer$by_var()

SOMAAxisIndexer$clone()

Method new()

Create a new SOMAAxisIndexer object.

Usage

SOMAAxisIndexer$new(query)

Arguments

query

The SOMAExperimentAxisQuery object to build indices for.

Method by_obs()

Get the index of the given obs coordinates.

Usage

SOMAAxisIndexer$by_obs(coords)

Arguments

coords

vector or arrow::Array of numeric coordinates.

Method by_var()

Get the index of the given var coordinates.

Usage

SOMAAxisIndexer$by_var(coords)

Arguments

coords

vector or arrow::Array of numeric coordinates.

Method clone()

The objects of this class are cloneable with this method.

Usage

SOMAAxisIndexer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.