SOMA Collection

Class for representing a SOMACollection, which may contain of one or more SOMAs.

Super classes

tiledbsoma::TileDBObject -> tiledbsoma::TileDBGroup -> SOMACollection

Public fields

uns

Named list of unstructured objects.

Active bindings

somas

Retrieve SOMA members.

Methods

Public methods

SOMACollection$new()

SOMACollection$set_query()

SOMACollection$reset_query()

SOMACollection$from_seurat()

SOMACollection$to_seurat()

SOMACollection$soma_uris()

SOMACollection$clone()

Inherited methods

tiledbsoma::TileDBObject$class()

tiledbsoma::TileDBObject$exists()

tiledbsoma::TileDBGroup$add_member()

tiledbsoma::TileDBGroup$add_metadata()

tiledbsoma::TileDBGroup$count_members()

tiledbsoma::TileDBGroup$get_member()

tiledbsoma::TileDBGroup$get_members()

tiledbsoma::TileDBGroup$get_metadata()

tiledbsoma::TileDBGroup$group_exists()

tiledbsoma::TileDBGroup$list_member_uris()

tiledbsoma::TileDBGroup$list_members()

tiledbsoma::TileDBGroup$list_object_uris()

tiledbsoma::TileDBGroup$list_objects()

tiledbsoma::TileDBGroup$print()

tiledbsoma::TileDBGroup$remove_member()

tiledbsoma::TileDBGroup$tiledb_group()

Method new()

Create a new SOMACollection. The existing TileDB group is opened at the specified array uri if one is present, otherwise a new array group is created. The members field is populated with SOMA objects for each URI passed explicitly to soma_uris, as well SOMA objects discovered within the SOMACollection object's TileDB group.

Usage

SOMACollection$new(uri, verbose = TRUE, config = NULL, ctx = NULL)

Arguments

uri

URI of the TileDB group

verbose

Print status messages

config

optional configuration

ctx

optional tiledb context

Method set_query()

Set query parameters to slice by dimension values or filter by attribute values that are applied to all SOMAs within the collection.

Usage

SOMACollection$set_query(
  obs_ids = NULL,
  var_ids = NULL,
  obs_attr_filter = NULL,
  var_attr_filter = NULL
)

Arguments

obs_ids, var_ids

character vector containing observation- or variable-identifiers.

obs_attr_filter, var_attr_filter

a TileDB query condition for attribute filtering pushdown.

Details

See SOMA$set_query() for more information about querying mechanics.

Method reset_query()

Reset query dimensions and attribute filters.

Usage

SOMACollection$reset_query()

Returns

NULL

Method from_seurat()

Convert a Seurat object to a TileDB-backed SOMACollection.

Assays

Each [SeuratObject::Assay] is converted to a SOMA and written to a nested TileDB group with a URI of ./soma_<assay> where <assay> is the name of the Seurat assay.

Identities

Cell identities in the SeuratObject::Seurat are maintained by creating an active_ident attribute in obs that stores the factor levels as a character vector.

Dimensionality Reductions

Dimensionality reduction results are stored as obsm and varm arrays within an SOMA. The SeuratObject::DimReduc object's key slot is used to determine which SOMA to store the results in. The array names are (obsm|varm)_dimreduction_<name>, where <name> is the name of the dimensionality reduction method (e.g., "pca").

Usage

SOMACollection$from_seurat(object)

Arguments

object

A SeuratObject::Seurat object.

Method to_seurat()

Convert to a SeuratObject::Seurat object.

Usage

SOMACollection$to_seurat(project = "SeuratProject", batch_mode = FALSE)

Arguments

project

SeuratObject::Project name for the Seurat object

batch_mode

logical, if TRUE, batch query mode is enabled for retrieving X layers. See AssayMatrix$to_dataframe() for more information.

Method soma_uris()

List the SOMA URIs in the collection.

Usage

SOMACollection$soma_uris()

Returns

A vector of URIs for each SOMA in the collection.

Method clone()

The objects of this class are cloneable with this method.

Usage

SOMACollection$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.