Base SOMA Context

R6 mapping class for SOMA context options. This class should be used as the basis for platform-specific contexts as it checks SOMA-specific context options

Super classes

tiledbsoma::MappingBase -> tiledbsoma::ScalarMap -> SOMAContextBase

Methods

Public methods

SOMAContextBase$new()

SOMAContextBase$set()

SOMAContextBase$clone()

Inherited methods

tiledbsoma::MappingBase$get()

tiledbsoma::MappingBase$items()

tiledbsoma::MappingBase$keys()

tiledbsoma::MappingBase$length()

tiledbsoma::MappingBase$print()

tiledbsoma::MappingBase$remove()

tiledbsoma::MappingBase$setv()

tiledbsoma::MappingBase$to_list()

tiledbsoma::MappingBase$update()

tiledbsoma::MappingBase$values()

Method new()

Usage

SOMAContextBase$new(config = NULL)

Arguments

config

...

Returns

This is a virtual class and cannot be directly instantiated

Method set()

Usage

SOMAContextBase$set(key, value)

Arguments

key

Key to set

value

Value to add for key, or NULL to remove the entry for key

Returns

[chainable] Invisibly returns self with value added for key

Method clone()

The objects of this class are cloneable with this method.

Usage

SOMAContextBase$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.