SOMAExperiment Axis Query

Perform an axis-based query against a SOMAExperiment.

SOMAExperimentAxisQuery allows easy selection and extraction of data from a single SOMAMeasurement in a SOMAExperiment, by obs/var (axis) coordinates and/or value filter. The primary use for this class is slicing SOMAExperiment X layers by obs or var value and/or coordinates. (lifecycle: experimental)

X Layer Support

Slicing on SOMASparseNDArray X matrices is supported; slicing on SOMADenseNDArray is not supported at this time.

Result Size

SOMAExperimentAxisQuery query class assumes it can store the full result of both axis dataframe queries in memory, and only provides incremental access to the underlying X NDArray. Accessors such as n_obs and n_vars codify this in the class.

Active bindings

experiment

The parent SOMAExperiment object.

indexer

The SOMAAxisIndexer object.

obs_query

The obs SOMAAxisQuery object.

var_query

The var SOMAAxisQuery object.

n_obs

The number of obs axis query results.

n_vars

The number of var axis query results.

obs_df

The obs SOMADataFrame object.

var_df

The var SOMADataFrame object for the specified measurement_name.

ms

The SOMAMeasurement object for the specified measurement_name.

Methods

Public methods

SOMAExperimentAxisQuery$new()

SOMAExperimentAxisQuery$obs()

SOMAExperimentAxisQuery$var()

SOMAExperimentAxisQuery$obs_joinids()

SOMAExperimentAxisQuery$var_joinids()

SOMAExperimentAxisQuery$X()

SOMAExperimentAxisQuery$read()

SOMAExperimentAxisQuery$to_seurat()

SOMAExperimentAxisQuery$to_seurat_assay()

SOMAExperimentAxisQuery$to_seurat_reduction()

SOMAExperimentAxisQuery$to_seurat_graph()

SOMAExperimentAxisQuery$clone()

Method new()

Create a new SOMAExperimentAxisQuery object.

Usage

SOMAExperimentAxisQuery$new(
  experiment,
  measurement_name,
  obs_query = NULL,
  var_query = NULL
)

Arguments

experiment

A SOMAExperiment object.

measurement_name

The name of the measurement to query.

obs_query, var_query

An SOMAAxisQuery object for the obs/var axis.

Method obs()

Retrieve obs arrow::Table

Usage

SOMAExperimentAxisQuery$obs(column_names = NULL)

Arguments

column_names

A character vector of column names to retrieve

Method var()

Retrieve var arrow::Table

Usage

SOMAExperimentAxisQuery$var(column_names = NULL)

Arguments

column_names

A character vector of column names to retrieve

Method obs_joinids()

Retrieve soma_joinids as an arrow::Array for obs.

Usage

SOMAExperimentAxisQuery$obs_joinids()

Method var_joinids()

Retrieve soma_joinids as an arrow::Array for var.

Usage

SOMAExperimentAxisQuery$var_joinids()

Method X()

Retrieves an X layer as an arrow::Table.

Usage

SOMAExperimentAxisQuery$X(layer_name)

Arguments

layer_name

The name of the layer to retrieve.

Method read()

Reads the entire query result as a list of arrow::Tables. This is a low-level routine intended to be used by loaders for other in-core formats, such as Seurat, which can be created from the resulting Tables.

Usage

SOMAExperimentAxisQuery$read(
  X_layers = NULL,
  obs_column_names = NULL,
  var_column_names = NULL
)

Arguments

X_layers

The name(s) of the X layer(s) to read and return.

obs_column_names, var_column_names

Specify which column names in var and obs dataframes to read and return.

Method to_seurat()

...

Usage

SOMAExperimentAxisQuery$to_seurat(
  X_layers = c(counts = "counts", data = "logcounts")
)

Arguments

X_layers

...

Returns

...

Method to_seurat_assay()

Loads the query as a Seurat Assay

Usage

SOMAExperimentAxisQuery$to_seurat_assay(
  X_layers = c(counts = "counts", data = "logcounts"),
  cells_index = NULL,
  features_index = NULL,
  var_column_names = NULL
)

Arguments

X_layers

A named character of X layers to add to the Seurat assay; names should be one of:

“counts” to add the layer as counts

“data” to add the layer as data

“scale.data” to add the layer as scale.data

At least one of “counts” or “data” is required

cells_index

Name of column in obs to add as cell names

features_index

Name of column in var to add as feature names

var_column_names

Names of columns in var to add as feature-level meta data

Returns

An Assay object

Method to_seurat_reduction()

Loads the query as a Seurat dimensional reduction

Usage

SOMAExperimentAxisQuery$to_seurat_reduction(
  embeddings,
  loadings = NULL,
  cells_index = NULL,
  features_index = NULL
)

Arguments

embeddings

Name of array in obsm to load as the cell embeddings

loadings

Name of the array in varm to load as the feature loadings; will try to determine loadings from embeddings

cells_index

Name of column in obs to add as cell names

features_index

Name of column in var to add as feature names

Returns

A DimReduc object

Method to_seurat_graph()

Loads the query as a Seurat graph

Usage

SOMAExperimentAxisQuery$to_seurat_graph(graph, cells_index = NULL)

Arguments

graph

Name of array in obsp to load as the graph

cells_index

Name of column in obs to add as cell names

Returns

A Graph object

Method clone()

The objects of this class are cloneable with this method.

Usage

SOMAExperimentAxisQuery$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.