SOMA TileDB Context

Context map for TileDB-backed SOMA objects

Super classes

tiledbsoma::MappingBase -> tiledbsoma::ScalarMap -> tiledbsoma::SOMAContextBase -> SOMATileDBContext

Methods

Public methods

SOMATileDBContext$new()

SOMATileDBContext$keys()

SOMATileDBContext$items()

SOMATileDBContext$length()

SOMATileDBContext$get()

SOMATileDBContext$set()

SOMATileDBContext$to_tiledb_context()

SOMATileDBContext$get_tiledb_context()

SOMATileDBContext$clone()

Inherited methods

tiledbsoma::MappingBase$print()

tiledbsoma::MappingBase$remove()

tiledbsoma::MappingBase$setv()

tiledbsoma::MappingBase$to_list()

tiledbsoma::MappingBase$update()

tiledbsoma::MappingBase$values()

Method new()

Usage

SOMATileDBContext$new(config = NULL, cached = TRUE)

Arguments

config

...

cached

Force new creation

Returns

An instantiated SOMATileDBContext object

Method keys()

Usage

SOMATileDBContext$keys()

Returns

The keys of the map

Method items()

Usage

SOMATileDBContext$items()

Returns

Return the items of the map as a list

Method length()

Usage

SOMATileDBContext$length()

Returns

The number of items in the map

Method get()

Usage

SOMATileDBContext$get(key, default = quote(expr = ))

Arguments

key

Key to fetch

default

Default value to fetch if key is not found; defaults to NULL

Returns

The value of key in the map, or default if key is not found

Method set()

Usage

SOMATileDBContext$set(key, value)

Arguments

key

Key to set

value

Value to add for key, or NULL to remove the entry for key

Returns

\[chainable\] Invisibly returns self with value added as key

Method to_tiledb_context()

Usage

SOMATileDBContext$to_tiledb_context()

Returns

A tiledb_ctx object, dynamically constructed. Most useful for the constructor of this class.

Method get_tiledb_context()

Usage

SOMATileDBContext$get_tiledb_context()

Returns

A tiledb_ctx object, which is a stored (and long-lived) result from to_tiledb_context.

Method clone()

The objects of this class are cloneable with this method.

Usage

SOMATileDBContext$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.