A Mapping Type for Scalars

An R6 mapping type that is limited to scalar atomic vector types only; can optionally be limited further to a specific atomic vector type (eg. “logical”).

Super class

tiledbsoma::MappingBase -> ScalarMap

Active bindings

type

The type that this ScalarMap is limited to

Methods

Public methods

ScalarMap$new()

ScalarMap$set()

ScalarMap$clone()

Inherited methods

tiledbsoma::MappingBase$get()

tiledbsoma::MappingBase$items()

tiledbsoma::MappingBase$keys()

tiledbsoma::MappingBase$length()

tiledbsoma::MappingBase$print()

tiledbsoma::MappingBase$remove()

tiledbsoma::MappingBase$setv()

tiledbsoma::MappingBase$to_list()

tiledbsoma::MappingBase$update()

tiledbsoma::MappingBase$values()

Method new()

Usage

ScalarMap$new(type = "any")

Arguments

type

Limit the ScalarMap to a preset type; choose from:

“any”

“numeric”

“integer”

“character”

“logical”

Returns

An instantiated ScalarMap object with the type set to type

Method set()

Usage

ScalarMap$set(key, value)

Arguments

key

Key to set

value

Value to add for key, or NULL to remove the entry for key

Returns

[chainable] Invisibly returns self with value added as key

Method clone()

The objects of this class are cloneable with this method.

Usage

ScalarMap$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.