SparseReadIter

SparseReadIter is a class that allows for iteration over a reads on SOMASparseNDArray. Iteration chunks are retrieved as 0-based Views matrixZeroBasedView of Matrix::sparseMatrix.

Super class

tiledbsoma::ReadIter -> SparseReadIter

Methods

Public methods

SparseReadIter$new()

SparseReadIter$concat()

SparseReadIter$clone()

Inherited methods

tiledbsoma::ReadIter$read_complete()

tiledbsoma::ReadIter$read_next()

Method new()

Create (lifecycle: maturing)

Usage

SparseReadIter$new(sr, shape, zero_based = FALSE)

Arguments

sr

Soma reader pointer

shape

Shape of the full matrix

zero_based

Logical, if TRUE will make iterator for Matrix::dgTMatrix-class otherwise matrixZeroBasedView.

Method concat()

Concatenate remainder of iterator.

Usage

SparseReadIter$concat()

Returns

matrixZeroBasedView of Matrix::sparseMatrix

Method clone()

The objects of this class are cloneable with this method.

Usage

SparseReadIter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.