TileDB Array Base Class

Base class for representing an individual TileDB array.

Details

Initialization

Initializing a TileDBArray object does not automatically create a new array at the specified uri if one does not already exist because we don't know what the schema will be. Arrays are only created by child classes, which populate the private create_empty_array() and ingest_data() methods.

Super class

tiledbsoma::TileDBObject -> TileDBArray

Methods

Public methods

TileDBArray$new()

TileDBArray$print()

TileDBArray$array_exists()

TileDBArray$tiledb_array()

TileDBArray$get_metadata()

TileDBArray$add_metadata()

TileDBArray$schema()

TileDBArray$dimensions()

TileDBArray$attributes()

TileDBArray$dimnames()

TileDBArray$fragment_count()

TileDBArray$attrnames()

TileDBArray$set_query()

TileDBArray$reset_query()

TileDBArray$clone()

Inherited methods

tiledbsoma::TileDBObject$class()

tiledbsoma::TileDBObject$exists()

Method new()

Create a new TileDBArray object.

Usage

TileDBArray$new(uri, verbose = TRUE, config = NULL, ctx = NULL)

Arguments

uri

URI for the TileDB array

verbose

Print status messages

config

optional configuration

ctx

optional tiledb context

Method print()

Print summary of the array.

Usage

TileDBArray$print()

Method array_exists()

Check if the array exists.

Usage

TileDBArray$array_exists()

Returns

TRUE if the array exists, FALSE otherwise.

Method tiledb_array()

Return a TileDBArray object

Usage

TileDBArray$tiledb_array(...)

Arguments

...

Optional arguments to pass to tiledb::tiledb_array()

Returns

A tiledb::tiledb_array object.

Method get_metadata()

Retrieve metadata from the TileDB array.

Usage

TileDBArray$get_metadata(key = NULL, prefix = NULL)

Arguments

key

The name of the metadata attribute to retrieve.

prefix

Filter metadata using an optional prefix. Ignored if key is not NULL.

Returns

A list of metadata values.

Method add_metadata()

Add list of metadata to the specified TileDB array.

Usage

TileDBArray$add_metadata(metadata, prefix = "")

Arguments

metadata

Named list of metadata to add.

prefix

Optional prefix to add to the metadata attribute names.

Returns

NULL

Method schema()

Retrieve the array schema

Usage

TileDBArray$schema()

Returns

A tiledb::tiledb_array_schema object

Method dimensions()

Retrieve the array dimensions

Usage

TileDBArray$dimensions()

Returns

A list of tiledb::tiledb_dim objects

Method attributes()

Retrieve the array attributes

Usage

TileDBArray$attributes()

Returns

A list of tiledb::tiledb_attr objects

Method dimnames()

Retrieve dimension names

Usage

TileDBArray$dimnames()

Returns

A character vector with the array's dimension names

Method fragment_count()

Get number of fragments in the array

Usage

TileDBArray$fragment_count()

Method attrnames()

Retrieve attribute names

Usage

TileDBArray$attrnames()

Returns

A character vector with the array's attribute names

Method set_query()

Set dimension values to slice from the array.

Usage

TileDBArray$set_query(dims = NULL, attr_filter = NULL)

Arguments

dims

a named list of character vectors. Each name must correspond to an array dimension. The character vectors within each element are used to set the arrays selected ranges for each corresponding dimension.

attr_filter

a TileDB query condition for attribute filtering. pushdown.

Method reset_query()

Reset the query. By default both dimension ranges and attribute filters are cleared.

Usage

TileDBArray$reset_query(dims = TRUE, attr_filter = TRUE)

Arguments

dims

Clear the defined dimension ranges?

attr_filter

Clear the defined attribute filters?

Returns

NULL

Method clone()

The objects of this class are cloneable with this method.

Usage

TileDBArray$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.