TileDBCreateOptions

Provides strongly-typed access and default values for platform_config options stored under the “tiledb” \(\rightarrow\) “create” mapping keys.

Intended for internal use only.

Super class

tiledbsoma::MappingBase -> TileDBCreateOptions

Methods

Public methods

TileDBCreateOptions$new()

TileDBCreateOptions$cell_tile_orders()

TileDBCreateOptions$dim_tile()

TileDBCreateOptions$capacity()

TileDBCreateOptions$allows_duplicates()

TileDBCreateOptions$dataframe_dim_zstd_level()

TileDBCreateOptions$sparse_nd_array_dim_zstd_level()

TileDBCreateOptions$offsets_filters()

TileDBCreateOptions$validity_filters()

TileDBCreateOptions$dim_filters()

TileDBCreateOptions$attr_filters()

TileDBCreateOptions$write_X_chunked()

TileDBCreateOptions$goal_chunk_nnz()

TileDBCreateOptions$.dim()

TileDBCreateOptions$.attr()

TileDBCreateOptions$.build_filters()

TileDBCreateOptions$.build_filter()

TileDBCreateOptions$clone()

Inherited methods

tiledbsoma::MappingBase$get()

tiledbsoma::MappingBase$items()

tiledbsoma::MappingBase$keys()

tiledbsoma::MappingBase$length()

tiledbsoma::MappingBase$print()

tiledbsoma::MappingBase$remove()

tiledbsoma::MappingBase$set()

tiledbsoma::MappingBase$setv()

tiledbsoma::MappingBase$to_list()

tiledbsoma::MappingBase$update()

tiledbsoma::MappingBase$values()

Method new()

Create a TileDBCreateOptions object

Usage

TileDBCreateOptions$new(platform_config)

Arguments

platform_config

A platform configuration object

Method cell_tile_orders()

Returns the cell and tile orders that should be used. If neither cell_order nor tile_order is present, only in this case will we use the default values provided.

Usage

TileDBCreateOptions$cell_tile_orders()

Returns

a list keyed by “cell_order” and “tile_order”, where either value or both may be NULL.

Method dim_tile()

Usage

TileDBCreateOptions$dim_tile(dim_name, default = .default_tile_extent())

Arguments

dim_name

Name of dimension to get tiling for

default

Default tiling if dim_name is not set

Returns

int

Method capacity()

Usage

TileDBCreateOptions$capacity()

Returns

int

Method allows_duplicates()

Usage

TileDBCreateOptions$allows_duplicates()

Returns

bool

Method dataframe_dim_zstd_level()

Usage

TileDBCreateOptions$dataframe_dim_zstd_level()

Returns

int

Method sparse_nd_array_dim_zstd_level()

Usage

TileDBCreateOptions$sparse_nd_array_dim_zstd_level()

Returns

int

Method offsets_filters()

Usage

TileDBCreateOptions$offsets_filters(default = .default_offsets_filters())

Arguments

default

Default offset filters to use if not currently set

Returns

list of tiledb.Filter

Method validity_filters()

Usage

TileDBCreateOptions$validity_filters(default = .default_validity_filters())

Arguments

default

Default validity filters to use if not currently set

Returns

list of tiledb.Filter

Method dim_filters()

Usage

TileDBCreateOptions$dim_filters(dim_name, default = list())

Arguments

dim_name

Name of dimension to get filters for

default

Default filters to use for if not currently set

Returns

list of tiledb.Filter

Method attr_filters()

Usage

TileDBCreateOptions$attr_filters(attr_name, default = list())

Arguments

attr_name

Name of attribute

default

Default filters to use if not currently set

Returns

list of tiledb.Filter

Method write_X_chunked()

Usage

TileDBCreateOptions$write_X_chunked()

Returns

bool

Method goal_chunk_nnz()

Usage

TileDBCreateOptions$goal_chunk_nnz()

Returns

int

Method .dim()

Usage

TileDBCreateOptions$.dim(dim_name)

Arguments

dim_name

Name of dimensions

Returns

Named list of character

Method .attr()

Usage

TileDBCreateOptions$.attr(attr_name)

Arguments

attr_name

Name of attribute

Returns

Named list of character

Method .build_filters()

Usage

TileDBCreateOptions$.build_filters(items)

Arguments

items

A list of filters; see $.build_filter() for details about entries in items

Returns

A list of tiledb filters.

Method .build_filter()

Usage

TileDBCreateOptions$.build_filter(item)

Arguments

item

THe name of a filter or a list with the name and arguments for a filter (eg. list(name = "ZSTD", COMPRESSION_LEVEL = -1))

Returns

A tiledb filter.

Method clone()

The objects of this class are cloneable with this method.

Usage

TileDBCreateOptions$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.