TileDB Group Base Class

Base class for interacting with TileDB groups

Value

A named list of group members.

The array object.

Details

Initialization

Upon initialization a new group is created if one does not already exist at the specified uri.

Super class

tiledbsoma::TileDBObject -> TileDBGroup

Public fields

members

Named list of members in the group

Methods

Public methods

TileDBGroup$new()

TileDBGroup$print()

TileDBGroup$group_exists()

TileDBGroup$tiledb_group()

TileDBGroup$list_objects()

TileDBGroup$list_object_uris()

TileDBGroup$add_member()

TileDBGroup$remove_member()

TileDBGroup$count_members()

TileDBGroup$list_members()

TileDBGroup$list_member_uris()

TileDBGroup$get_members()

TileDBGroup$get_member()

TileDBGroup$get_metadata()

TileDBGroup$add_metadata()

TileDBGroup$clone()

Inherited methods

tiledbsoma::TileDBObject$class()

tiledbsoma::TileDBObject$exists()

Method new()

Create a new TileDBGroup object.

Usage

TileDBGroup$new(uri, verbose = TRUE, config = NULL, ctx = NULL)

Arguments

uri

TileDB array URI

verbose

Print status messages

config

optional configuration

ctx

optional tiledb context

Method print()

Print summary of the group.

Usage

TileDBGroup$print()

Method group_exists()

Check if the group exists.

Usage

TileDBGroup$group_exists()

Returns

TRUE if the group exists, FALSE otherwise.

Method tiledb_group()

Return a tiledb_group object

Usage

TileDBGroup$tiledb_group(...)

Arguments

...

Optional arguments to pass to tiledb::tiledb_array()

Returns

A tiledb::tiledb_group object.

Method list_objects()

List the TileDB objects within the group.

Usage

TileDBGroup$list_objects(type = NULL)

Arguments

type

The type of object to list, either "ARRAY", or "GROUP". By default all object types are listed.

Returns

A data.frame with columns URI and TYPE.

Method list_object_uris()

List URIs for TileDB objects within the group.

Usage

TileDBGroup$list_object_uris(type = NULL, prefix = NULL)

Arguments

type

The type of object to list, either "ARRAY", or "GROUP". By default all object types are listed.

prefix

Filter URIs whose basename contain an optional prefix.

Returns

A character vector of object URIs with names corresponding to the basename of the object.

Method add_member()

Add new member to the group.

Usage

TileDBGroup$add_member(object, name = NULL, relative = NULL)

Arguments

object

The TileDBArray or TileDBGroup object to add.

name

The name to use for the member. By default the base name of the object's URI is used.

relative

A logical value indicating whether the new member's URI is relative to the group's URI.

Method remove_member()

Remove member from the group.

Usage

TileDBGroup$remove_member(name)

Arguments

name

The name of the member to remove.

Method count_members()

Count the number of members in the group.

Usage

TileDBGroup$count_members()

Returns

Integer count of members in the group.

Method list_members()

List the members of the group.

Usage

TileDBGroup$list_members(type = NULL)

Arguments

type

The type of member to list, either "ARRAY", or "GROUP". By default all member types are listed.

Returns

A data.frame with columns URI, TYPE, and NAME.

Method list_member_uris()

List URIs for group members

Usage

TileDBGroup$list_member_uris(type = NULL, prefix = NULL)

Arguments

type

The type of member to list, either "ARRAY", or "GROUP". By default all member types are listed.

prefix

Filter for members whose name contains an optional prefix.

Returns

A character vector of member URIs, named for the group member

Method get_members()

Retrieve arrays within the group that meet the specified criteria.

Usage

TileDBGroup$get_members(type = NULL, prefix = NULL)

Arguments

type

The type of group members to list, either "ARRAY", or "GROUP".

prefix

String prefix to filter the member names.

Method get_member()

Retrieve a group member.

Usage

TileDBGroup$get_member(name)

Arguments

name

The name of the array to retrieve.

Method get_metadata()

Retrieve metadata from the TileDB group.

Usage

TileDBGroup$get_metadata(key = NULL, prefix = NULL)

Arguments

key

The name of the metadata attribute to retrieve.

prefix

Filter metadata using an optional prefix. Ignored if key is not NULL.

Returns

A list of metadata values.

Method add_metadata()

Add list of metadata to the TileDB group.

Usage

TileDBGroup$add_metadata(metadata, prefix = "")

Arguments

metadata

Named list of metadata to add.

prefix

Optional prefix to add to the metadata attribute names.

Returns

NULL

Method clone()

The objects of this class are cloneable with this method.

Usage

TileDBGroup$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.