TileDB Object Base Class

Base class to implement shared functionality across the TileDBArray and TileDBGroup classes. (lifecycle: experimental)

Active bindings

platform_config

Platform configuration

tiledbsoma_ctx

SOMATileDBContext

uri

The URI of the TileDB object.

Methods

Public methods

TileDBObject$new()

TileDBObject$class()

TileDBObject$is_open()

TileDBObject$mode()

TileDBObject$print()

TileDBObject$exists()

TileDBObject$clone()

Method new()

Create a new TileDB object. (lifecycle: experimental)

Usage

TileDBObject$new(
  uri,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  tiledb_timestamp = NULL,
  internal_use_only = NULL
)

Arguments

uri

URI for the TileDB object

platform_config

Optional platform configuration

tiledbsoma_ctx

Optional SOMATileDBContext

tiledb_timestamp

Optional Datetime (POSIXct) with TileDB timestamp

internal_use_only

Character value to signal this is a 'permitted' call, as new() is considered internal and should not be called directly.

Method class()

Print the name of the R6 class.

Usage

TileDBObject$class()

Method is_open()

Determine if the object is open for reading or writing

Usage

TileDBObject$is_open()

Returns

TRUE if the object is open, otherwise FALSE

Method mode()

Get the mode of the object

Usage

TileDBObject$mode()

Returns

If the object is closed, returns “CLOSED”; otherwise returns the mode (eg. “READ”) of the object

Method print()

Print-friendly representation of the object.

Usage

TileDBObject$print()

Method exists()

Check if the object exists. (lifecycle: experimental)

Usage

TileDBObject$exists()

Returns

TRUE`` if the object exists, FALSE` otherwise.

Method clone()

The objects of this class are cloneable with this method.

Usage

TileDBObject$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.