Get the R Type from an Arrow Type

Get an R type from an Arrow type. This function is equivalent to typeof() rather than mode() or class(), and returns the equivalent type. For example, the equivalent R type to an Arrow dictionary is “integer”, not “factor”; likewise, the equivalent R type to an Arrow 64-bit integer is “double”

Usage

r_type_from_arrow_type(x)

# S3 method for class 'Schema'
r_type_from_arrow_type(x)

# S3 method for class 'Field'
r_type_from_arrow_type(x)

# S3 method for class 'DataType'
r_type_from_arrow_type(x)

Arguments

x

An Arrow schema, field, or data type

Value

If x is a data type, a single character value giving the R type of x; if no corresponding R type, returns the Arrow type name

If x is a field, a single named character vector with the name being the field name and the value being the R type

If x is a schema, a named vector where the names are field names and the values are the R types of each field

See also