Read SOMA Data From a Given URI

This functions access a given SOMA URI and returns a complete data.frame. It does not iterate; if your data is large than the initial read size consider the sr_* functions.

soma_array_reader(
  uri,
  colnames = NULL,
  qc = NULL,
  dim_points = NULL,
  dim_ranges = NULL,
  batch_size = "auto",
  result_order = "auto",
  loglevel = "auto",
  config = NULL
)

nnz(uri, config = NULL)

shape(uri, config = NULL)

Arguments

uri

Character value with URI path to a SOMA data set

colnames

Optional vector of character value with the name of the columns to retrieve

qc

Optional external Pointer object to TileDB Query Condition, defaults to ‘NULL’ i.e. no query condition

dim_points

Optional named list with vector of data points to select on the given dimension(s). Each dimension can be one entry in the list.

dim_ranges

Optional named list with two-column matrix where each row select a range for the given dimension. Each dimension can be one entry in the list.

batch_size

Character value with the desired batch size, defaults to ‘auto’

result_order

Character value with the desired result order, defaults to ‘auto’

loglevel

Character value with the desired logging level, defaults to ‘auto’ which lets prior setting prevail, any other value is set as new logging level.

config

Optional character vector containing TileDB config.

Value

A List object with two pointers to Arrow array data and schema is returned

Examples

if (FALSE) {
uri <- "test/soco/pbmc3k_processed/obs"
z <- soma_array_reader(uri)
tb <- as_arrow_table(z)
}