Write R Objects to SOMA

Various helpers to write R objects to SOMA

# S3 method for data.frame
write_soma(
  x,
  uri,
  soma_parent,
  df_index = NULL,
  index_column_names = "soma_joinid",
  ...,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

# S3 method for matrix
write_soma(
  x,
  uri,
  soma_parent,
  sparse = TRUE,
  type = NULL,
  transpose = FALSE,
  ...,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

# S3 method for Matrix
write_soma(
  x,
  uri,
  soma_parent,
  sparse = TRUE,
  type = NULL,
  transpose = FALSE,
  ...,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

# S3 method for TsparseMatrix
write_soma(
  x,
  uri,
  soma_parent,
  type = NULL,
  transpose = FALSE,
  ...,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

Arguments

x

An object

uri

URI for resulting SOMA object

soma_parent

The parent collection (eg. a SOMACollection, SOMAExperiment, or SOMAMeasurement)

df_index

The name of the column in x with the index (row names); by default, will automatically add the row names of x to an attribute named “index” to the resulting SOMADataFrame

index_column_names

Names of columns in x to index in the resulting SOMA object

...

Arguments passed to other methods

platform_config

Optional platform configuration

tiledbsoma_ctx

Optional SOMATileDBContext

relative

[Internal use only] Is uri relative or aboslute

sparse

Create a sparse or dense array from x

type

Arrow type for encoding x (eg. arrow::int32()); by default, attempts to determine arrow type with arrow::infer_type()

transpose

Transpose x before writing

Value

The resulting SOMA array or data frame

Writing Data Frames

Data frames are written out as SOMADataFrames. The following transformations are applied to x:

row names are added to a column in x entitled “index”, “_index”, or a random name if either option is already present in x

a column “soma_joinid” will be automatically added going from [0, nrow(x) - 1] encoded as 64-bit integers

all factor columns will be coerced to characters

The array type for each column will be determined by arrow::infer_type(); if any column contains a non-atomic type (excluding factors, complexes,and raws), the code will error out

Writing Dense Matrices

Dense matrices are written as two-dimensional dense arrays. The overall shape of the array is determined by dim(x) and the type of the array is determined by type or arrow::infer_type(x)

Writing Sparse Matrices

Sparse matrices are written out as two-dimensional TileDB sparse arrays in COO format:

the row indices (“i”) are written out as “soma_dim_0”

the column indices (“j”) are written out as “soma_dim_1”

the non-zero values (“x”) are written out as “soma_data”

The array type is determined by type, or arrow::infer_type(slot(x, "x"))