Convert a Seurat Sub-Object to a SOMA Object, returned opened for write

Various helpers to write Seurat sub-objects to SOMA objects..

Usage

# S3 method for class 'Assay'
write_soma(
  x,
  uri = NULL,
  soma_parent,
  ...,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  context = NULL,
  relative = TRUE
)

# S3 method for class 'Assay5'
write_soma(
  x,
  uri = NULL,
  soma_parent,
  ...,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  context = NULL,
  relative = TRUE
)

# S3 method for class 'DimReduc'
write_soma(
  x,
  uri = NULL,
  soma_parent,
  fidx = NULL,
  nfeatures = NULL,
  ...,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  context = NULL,
  relative = TRUE
)

# S3 method for class 'Graph'
write_soma(
  x,
  uri,
  soma_parent,
  ...,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  context = NULL,
  relative = TRUE
)

# S3 method for class 'SeuratCommand'
write_soma(
  x,
  uri = NULL,
  soma_parent,
  ...,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  context = NULL,
  relative = TRUE
)

Arguments

x

A Seurat sub-object (eg. an Assay, a DimReduc, or a Graph).

uri

URI for resulting SOMA object.

soma_parent

The parent collection; for the DimReduc and Graph methods, this must be a measurement for the assay x was generated from.

...

Arguments passed to other methods

ingest_mode

Ingestion mode when creating the SOMA; choose from:

“write”: create a new SOMA and error if it already exists.

“resume”: attempt to create a new SOMA; if it already exists, simply open it for writing.

platform_config

Optional platform configuration.

tiledbsoma_ctx

Optional (DEPRECATED) SOMATileDBContext.

context

Optional SOMAContext object used for TileDB operations. If a context is not provided, then the default context will be used. Call set_default_context once before other SOMA operations to configure the default context.

relative

[Internal use only] Is uri relative or absolute.

fidx

An integer vector describing the location of features in SeuratObject::Loadings(x) with relation to soma_parent (eg. match(rownames(Loadings(x)), rownames(assay))).

nfeatures

The number of features present in soma_parent (eg. nrow(assay)).

Value

Assay and Assay5 methods: a SOMAMeasurement with the data from x, returned opened for write.

DimReduc and Graph methods: invisibly returns soma_parent, opened for write, with the values of x added to it.

Writing v3 Assays

Seurat Assay objects are written out as individual measurements:

the “data” matrix is written out as a sparse array called “data” within the “X” group.

the “counts” matrix, if not empty, is written out as a sparse array called “counts” within the “X” group.

the “scale.data” matrix, if not empty, is written out as a sparse array called “scale_data” within the “X” group.

feature-level metadata is written out as a data frame called “var”.

Expression matrices are transposed (cells as rows) prior to writing. All other slots, including results from extended assays (eg. SCTAssay, ChromatinAssay) are lost.

Performance Considerations

Ingestion of very large dense layers, such as scale.data, can be memory intensive. For better performance, users can remove these layers prior to ingestion and regenerate them after export, or ingest them separately as dense arrays for those who need to persist the exact matrix

# Using SeuratObject v5 syntax on a v3 `Assay`
# Cache the layer for separate ingestion, skip if planning to regenerate
mat <- object[["ASSAY"]]$scale.data

# Remove the `scale.data` layer
object[["ASSAY"]]$scale.data <- NULL

# Ingest the smaller object
uri <- write_soma(object, "/path/to/soma")

# Ingest the `scale.data` layer densely; needed only if persistence
# of the data is paramount
# Pad the `scale.data` layer so that its soma join IDs match the experiment
padded <- matrix(
  data = vector("numeric", length = prod(dim(object[["ASSAY"]]))),
  nrow = nrow(object[["ASSAY"]]),
  ncol = ncol(object[["ASSAY"]])
)
rowidx <- match(rownames(mat), rownames(object[["ASSAY"]]))
colidx <- match(colnames(mat), colnames(object[["ASSAY"]]))
padded[rowidx, colidx] <- mat

# Use `write_soma()` to ingest densely and register it within the `uns`
# collection; this may need to be created manually if the original
# object does not contain command logs
exp <- SOMAExperimentOpen(uri, "WRITE")
if (!match("uns", exp$names(), nomatch = 0L)) {
  # For `tiledb://` URIs, set the URI for the new collection manually rather
  # than relying on `file.path()`
  uns <- SOMACollectionCreate(file.path(exp$uri, "uns"))
  exp$add_new_collection(uns, "uns")
}
arr <- write_soma(
  padded,
  "scale_data",
  soma_parent = exp$get("uns"),
  sparse = FALSE,
  key = "scale_data"
)
arr$close()
exp$close()

Please note that dense arrays cannot be read in using the SOMAExperimentAxisQuery mechanism; use SOMADenseNDArray$read_dense_matrix, remembering to transpose before adding back to a Seurat object

Writing v5 Assays

Seurat v5 Assayss are written out as individual measurements:

the layer matrices are written out as sparse arrays within the “X” group.

feature-level metadata is written out as a data frame called “var”.

Expression matrices are transposed (cells as rows) prior to writing. All other slots, including results from extended assays (eg. SCTAssay, ChromatinAssay) are lost.
The following bits of metadata are written in various parts of the measurement

“soma_ecosystem_seurat_assay_version”: written at the measurement level; indicates the Seurat assay version. Set to “v5”.

“soma_ecosystem_seurat_v5_default_layers”: written at the “X” group level; indicates the default layers.

“soma_ecosystem_seurat_v5_ragged”: written at the “X/<layer>” array level; with a value of “ragged”, indicates whether or not the layer is ragged.

“soma_r_type_hint”: written at the “X/<layer>” array level; indicates the R class and defining package (for S4 classes) of the original layer.

Writing DimReducs

Seurat DimReduc objects are written out to the “obsm” and “varm” groups of a measurement:

cell embeddings are written out as a sparse matrix in the “obsm” group.

feature loadings, if not empty, are written out as a sparse matrix in the “varm” groups; loadings are padded with NAs to include all features.

Dimensional reduction names are translated to AnnData-style names (eg. “pca” becomes X_pca for embeddings and “PCs” for loadings). All other slots, including projected feature loadings and jackstraw information, are lost.

Writing Graphs

Seurat Graph objects are written out as sparse matrices to the “obsp” group of a measurement.

Writing SeuratCommands

Seurat command logs are written out as data frames to the “seurat_commands” group of a collection.

Examples

uri <- withr::local_tempfile(pattern = "seurat-sub")

data("pbmc_small", package = "SeuratObject")
suppressWarnings(pbmc_small <- SeuratObject::UpdateSeuratObject(pbmc_small))
#> Validating object structure
#> Updating object slots
#> Ensuring keys are in the proper structure
#> Updating matrix keys for DimReduc ‘pca’
#> Updating matrix keys for DimReduc ‘tsne’
#> Ensuring keys are in the proper structure
#> Ensuring feature names don't have underscores or pipes
#> Updating slots in RNA
#> Updating slots in RNA_snn
#> Setting default assay of RNA_snn to RNA
#> Updating slots in pca
#> Updating slots in tsne
#> Setting tsne DimReduc to global
#> Setting assay used for NormalizeData.RNA to RNA
#> Setting assay used for ScaleData.RNA to RNA
#> Setting assay used for RunPCA.RNA to RNA
#> Setting assay used for BuildSNN.RNA.pca to RNA
#> No assay information could be found for FindClusters
#> Setting assay used for RunTSNE.pca to RNA
#> Setting assay used for JackStraw.RNA.pca to RNA
#> Setting assay used for ScoreJackStraw.pca to RNA
#> Setting assay used for ProjectDim.RNA.pca to RNA
#> Setting assay used for FindVariableFeatures.RNA to RNA
#> Validating object structure for Assay ‘RNA’
#> Validating object structure for Graph ‘RNA_snn’
#> Validating object structure for DimReduc ‘pca’
#> Validating object structure for DimReduc ‘tsne’
#> Object representation is consistent with the most current Seurat version

col <- SOMACollectionCreate(uri)

# Write a v3 Assay
(assay <- pbmc_small[["RNA"]])
#> Assay data with 230 features for 80 cells
#> Top 10 variable features:
#>  PPBP, IGLL5, VDAC3, CD1C, AKR1C3, PF4, MYL9, GNLY, TREML1, CA2 
(ms <- write_soma(assay, "RNA", soma_parent = col))
#> <SOMAMeasurement>
#>   uri: /tmp/Rtmpfr8mYm/seurat-sub2a7e7537a170/RNA
#>   arrays: var 
#>   groups: X 

# Write a v5 Assay
(assay5 <- methods::as(pbmc_small[["RNA"]], "Assay5"))
#> Assay (v5) data with 230 features for 80 cells
#> Top 10 variable features:
#>  PPBP, IGLL5, VDAC3, CD1C, AKR1C3, PF4, MYL9, GNLY, TREML1, CA2 
#> Layers:
#>  data, counts, scale.data 
(ms5 <- write_soma(assay5, "RNA5", soma_parent = col))
#> <SOMAMeasurement>
#>   uri: /tmp/Rtmpfr8mYm/seurat-sub2a7e7537a170/RNA5
#>   arrays: var 
#>   groups: X 

ms5$close()

# Write a dimensional reduction
(tsne <- pbmc_small[["tsne"]])
write_soma(tsne, soma_parent = ms)
ms$obsm
#> <SOMACollection>
#>   uri: /tmp/Rtmpfr8mYm/seurat-sub2a7e7537a170/RNA/obsm
#>   arrays: X_tsne 

# Write a Seurat Graph
(snn <- pbmc_small[["RNA_snn"]])
#> A Graph object containing 80 cells
write_soma(snn, "snn", soma_parent = ms)
ms$obsp
#> <SOMACollection>
#>   uri: /tmp/Rtmpfr8mYm/seurat-sub2a7e7537a170/RNA/obsp
#>   arrays: snn 

# Write a Seurat command log
(cmd <- pbmc_small[["NormalizeData.RNA"]])
#> Command: NormalizeData(object = pbmc_small)
#> Time: 2018-08-27 20:32:17.136699
#> assay : RNA 
#> normalization.method : LogNormalize 
#> scale.factor : 10000 
#> verbose : TRUE 
write_soma(cmd, "NormalizeData.RNA", soma_parent = col)
(logs <- col$get("seurat_commands"))
#> <SOMACollection>
#>   uri: /tmp/Rtmpfr8mYm/seurat-sub2a7e7537a170/seurat_commands
#>   arrays: NormalizeData.RNA 
logs$get("NormalizeData.RNA")
#> <SOMADataFrame>
#>   uri: /tmp/Rtmpfr8mYm/seurat-sub2a7e7537a170/seurat_commands/NormalizeData.RNA
#>   dimensions: soma_joinid 
#>   attributes: values 

col$close()