A Configuration List

An R6 mapping type for configuring various “parameters”. Essentially, serves as a nested map where the inner map is a ScalarMap: {<param>: {<key>: <value>}}

Super class

tiledbsoma::MappingBase -> ConfigList

Methods

Public methods

ConfigList$get()

ConfigList$set()

ConfigList$setv()

ConfigList$clone()

Inherited methods

tiledbsoma::MappingBase$initialize()

tiledbsoma::MappingBase$items()

tiledbsoma::MappingBase$keys()

tiledbsoma::MappingBase$length()

tiledbsoma::MappingBase$print()

tiledbsoma::MappingBase$remove()

tiledbsoma::MappingBase$to_list()

tiledbsoma::MappingBase$update()

tiledbsoma::MappingBase$values()

Method get()

Usage

ConfigList$get(param, key = NULL, default = quote(expr = ))

Arguments

param

Outer key or “parameter” to fetch

key

Inner key to fetch; pass NULL to return the map for param

default

Default value to fetch if key is not found; defaults to NULL

Returns

The value of key for param in the map, or default if key is not found

Method set()

Usage

ConfigList$set(param, key, value)

Arguments

param

Outer key or “parameter” to set

key

Inner key to set

value

Value to add for key, or NULL to remove the entry for key; optionally provide only param and value as a ScalarMap to update param with the keys and values from value

Returns

\[chainable\] Invisibly returns self with value added for key in param

Method setv()

Usage

ConfigList$setv(...)

Arguments

...

Ignored

Returns

Nothing; setv() is disabled for ConfigList objects

Method clone()

The objects of this class are cloneable with this method.

Usage

ConfigList$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.