The Coordinate Strider

The CoordsStrider allows creating coordinate slices in an interated manner. Alternatively, it can chunk an existing vector of coordinates

Note

The CoordsStrider operates using 64-bit integer objects; as such, accessing fields, such as strider$start or strider$stride will return an integer64 object, which functions differently than a regular integer. Use with caution and convert back to integers or numerics as necessary

Active bindings

coords

If set, the coordinates to iterate over

start

If set, the starting point of the iterated coordinates; otherwise the minimum value of self$coords

end

If set, the end point of the iterated coordinates; otherwise the maximum value of self$coords

stride

The stride, or how many coordinates to generate per iteration; note: this field is settable, which will reset the iterator

Methods

Public methods

CoordsStrider$new()

CoordsStrider$print()

CoordsStrider$length()

CoordsStrider$has_next()

CoordsStrider$next_element()

Method new()

Create a coordinate strider

Usage

CoordsStrider$new(coords, ..., stride = NULL, start = NULL, end = NULL)

Arguments

coords

An integer vector of coordinates

...

Ignored

stride

The stride of how many coordinates to yield per iteration; by default, will try to yield all coordinates per iteration

start

If coords is missing, the starting coordinate to generate

end

If coords is missing, the ending coordinate to generate

Method print()

Print the coordinate strider to the screen

Usage

CoordsStrider$print()

Method length()

Get the length (span) of the coordinates within the strider

Usage

CoordsStrider$length()

Returns

The length (span) of the coordinate strider

Method has_next()

Determine if there are more coordinates to yield

Usage

CoordsStrider$has_next()

Returns

TRUE if there are more coordinates to yield or FALSE if otherwise

Method next_element()

Generate the next set of coordinates to yield. If there are no more coordinates to yield, raises a stopIteration error

Usage

CoordsStrider$next_element()

Returns

An integer vector of the next set of coordinates

Examples

strider <- CoordsStrider$new(start = 1L, end = 200L, stride = 60L)
while (strider$has_next()) {
  str(strider$next_element())
}
#> integer64 [1:60] 1 2 3 4 5 6 7 8 ... 
#> integer64 [1:60] 61 62 63 64 65 66 67 68 ... 
#> integer64 [1:60] 121 122 123 124 125 126 127 128 ... 
#> integer64 [1:20] 181 182 183 184 185 186 187 188 ...