Ephemeral SOMA Measurement

Ephemeral version of SOMAMeasurements; ephemeral measurements are equivalent to SOMA measurements but are stored in-memory instead of on-disk

Super classes

tiledbsoma::TileDBObject -> tiledbsoma::TileDBGroup -> tiledbsoma::SOMACollectionBase -> tiledbsoma::EphemeralCollectionBase -> EphemeralMeasurement

Active bindings

var

A SOMADataFrame containing primary annotations on the variable axis, for variables in this measurement (i.e., annotates columns of X). The contents of the soma_joinid column define the variable index domain, var_id. All variables for this measurement must be defined in this data frame

X

A SOMACollection of SOMASparseNDArrays; each contain measured feature values indexed by [obsid, varid]

obsm

A SOMACollection of SOMADenseNDArrays containing annotations on the observation axis. Each array is indexed by obsid and has the same shape as obs

obsp

A SOMACollection of SOMASparseNDArrays containing pairwise annotations on the observation axis and indexed with [obsid_1, obsid_2]

varm

A SOMACollection of SOMADenseNDArrays containing annotations on the variable axis. Each array is indexed by varid and has the same shape as var

varp

A SOMACollection of SOMASparseNDArrays containing pairwise annotations on the variable axis and indexed with [varid_1, varid_2]

soma_type

The SOMA object type

Methods

Public methods

EphemeralMeasurement$clone()

Inherited methods

tiledbsoma::TileDBObject$class()

tiledbsoma::TileDBObject$is_open()

tiledbsoma::TileDBObject$mode()

tiledbsoma::TileDBObject$reopen()

tiledbsoma::TileDBGroup$to_data_frame()

tiledbsoma::TileDBGroup$to_list()

tiledbsoma::EphemeralCollectionBase$add_new_collection()

tiledbsoma::EphemeralCollectionBase$add_new_dataframe()

tiledbsoma::EphemeralCollectionBase$add_new_dense_ndarray()

tiledbsoma::EphemeralCollectionBase$add_new_sparse_ndarray()

tiledbsoma::EphemeralCollectionBase$close()

tiledbsoma::EphemeralCollectionBase$create()

tiledbsoma::EphemeralCollectionBase$exists()

tiledbsoma::EphemeralCollectionBase$get()

tiledbsoma::EphemeralCollectionBase$get_metadata()

tiledbsoma::EphemeralCollectionBase$get_tiledb_config()

tiledbsoma::EphemeralCollectionBase$initialize()

tiledbsoma::EphemeralCollectionBase$length()

tiledbsoma::EphemeralCollectionBase$names()

tiledbsoma::EphemeralCollectionBase$open()

tiledbsoma::EphemeralCollectionBase$print()

tiledbsoma::EphemeralCollectionBase$remove()

tiledbsoma::EphemeralCollectionBase$set()

tiledbsoma::EphemeralCollectionBase$set_metadata()

Method clone()

The objects of this class are cloneable with this method.

Usage

EphemeralMeasurement$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.