Platform Configuration

An R6 mapping type for configuring various “parameters” for multiple “platforms”, essentially serves a multi-nested map where the inner map is a ScalarMap contained within a ConfigList (middle map): {platform: {param: {key: value}}}

Super class

tiledbsoma::MappingBase -> PlatformConfig

Methods

Public methods

PlatformConfig$platforms()

PlatformConfig$params()

PlatformConfig$get()

PlatformConfig$get_params()

PlatformConfig$set()

PlatformConfig$setv()

PlatformConfig$clone()

Inherited methods

tiledbsoma::MappingBase$initialize()

tiledbsoma::MappingBase$items()

tiledbsoma::MappingBase$keys()

tiledbsoma::MappingBase$length()

tiledbsoma::MappingBase$print()

tiledbsoma::MappingBase$remove()

tiledbsoma::MappingBase$to_list()

tiledbsoma::MappingBase$update()

tiledbsoma::MappingBase$values()

Method platforms()

Usage

PlatformConfig$platforms()

Returns

The names of the “platforms” (outer keys)

Method params()

Usage

PlatformConfig$params(platform = NULL)

Arguments

platform

The “platform” to pull parameter names (middle keys) for; pass TRUE to return all possible parameter names

Returns

The parameter names (middle keys) for platform

Method get()

Usage

PlatformConfig$get(
  platform,
  param = NULL,
  key = NULL,
  default = quote(expr = )
)

Arguments

platform

The name of the “platform” (outer key) to fetch

param

The name of the “paramters” of platform to fetch; if NULL, returns the configuration for platform

key

The “key” (inner key) for param in platform to fetch; if NULL and param is passed, returns the map for param in platform

default

Default value to fetch if key is not found; defaults to null

Returns

The value of key for param in platform in the map, or default if key is not found

Method get_params()

Usage

PlatformConfig$get_params(platform)

Arguments

platform

The name of the “platform” (outer key) to fetch

Returns

The ConfigList for platform

Method set()

Usage

PlatformConfig$set(platform, param, key, value)

Arguments

platform

The name of the “platform” (outer key) to set

param

Name of the “parameter” (middle key) in platform to set

key

Inner key to set

value

Value to add for key, or NULL to remove the entry for key; optionally provide only platfomr, param, and value as a ScalarMap to update param for platform with the keys and values from value

Returns

\[chainable\] Invisibly returns self with value added for key in param for platform

Method setv()

Usage

PlatformConfig$setv(...)

Arguments

...

Ignored

Returns

Nothing; setv() is disabled for PlatformConfig objects

Method clone()

The objects of this class are cloneable with this method.

Usage

PlatformConfig$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.