SOMA Collection Base Class

Base class for objects containing persistent collection of SOMA objects, mapping string keys to any SOMA object. (lifecycle: experimental)

Value

SOMA object.

Super classes

tiledbsoma::TileDBObject -> tiledbsoma::TileDBGroup -> SOMACollectionBase

Active bindings

soma_type

Retrieve the SOMA object type.

Methods

Public methods

SOMACollectionBase$new()

SOMACollectionBase$create()

SOMACollectionBase$set()

SOMACollectionBase$get()

SOMACollectionBase$add_new_collection()

SOMACollectionBase$add_new_dataframe()

SOMACollectionBase$add_new_dense_ndarray()

SOMACollectionBase$add_new_sparse_ndarray()

SOMACollectionBase$clone()

Inherited methods

tiledbsoma::TileDBObject$class()

tiledbsoma::TileDBObject$exists()

tiledbsoma::TileDBObject$is_open()

tiledbsoma::TileDBObject$mode()

tiledbsoma::TileDBGroup$close()

tiledbsoma::TileDBGroup$get_metadata()

tiledbsoma::TileDBGroup$length()

tiledbsoma::TileDBGroup$names()

tiledbsoma::TileDBGroup$open()

tiledbsoma::TileDBGroup$print()

tiledbsoma::TileDBGroup$remove()

tiledbsoma::TileDBGroup$set_metadata()

tiledbsoma::TileDBGroup$to_data_frame()

tiledbsoma::TileDBGroup$to_list()

Method new()

Create a new SOMACollection. (lifecycle: experimental)

Usage

SOMACollectionBase$new(
  uri,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  internal_use_only = NULL
)

Arguments

uri

URI of the TileDB group

platform_config

Optional storage-engine specific configuration

tiledbsoma_ctx

optional SOMATileDBContext

internal_use_only

Character value to signal this is a 'permitted' call, as new() is considered internal and should not be called directly.

Method create()

Add a new SOMA object to the collection. (lifecycle: experimental)

Usage

SOMACollectionBase$create(internal_use_only = NULL)

Arguments

internal_use_only

Character value to signal this is a 'permitted' call, as create() is considered internal and should not be called directly.

Method set()

Add a new SOMA object to the collection. (lifecycle: experimental)

Usage

SOMACollectionBase$set(object, name = NULL, relative = NULL)

Arguments

object

SOMA object.

name

The name to use for the object. Defaults to the object URI's base name.

relative

An optional logical value indicating whether the new object's URI is relative to the collection's URI. If NULL (the default), the object's URI is assumed to be relative unless it is a tiledb:// URI.

Method get()

Retrieve a SOMA object by name. (lifecycle: experimental)

Usage

SOMACollectionBase$get(name, mode = "READ")

Arguments

name

The name of the object to retrieve.

mode

Mode to open in

Method add_new_collection()

Add a new SOMA collection to this collection. (lifecycle: experimental)

Usage

SOMACollectionBase$add_new_collection(object, key)

Arguments

object

SOMA collection object.

key

The key to be added.

Method add_new_dataframe()

Add a new SOMA dataframe to this collection. (lifecycle: experimental)

Usage

SOMACollectionBase$add_new_dataframe(key, schema, index_column_names)

Arguments

key

The key to be added.

schema

Arrow schema argument passed on to DataFrame$create()

index_column_names

Index column names passed on to DataFrame$create()

Method add_new_dense_ndarray()

Add a new SOMA DenseNdArray to this collection. (lifecycle: experimental)

Usage

SOMACollectionBase$add_new_dense_ndarray(key, type, shape)

Arguments

key

The key to be added.

type

an Arrow type defining the type of each element in the array.

shape

a vector of integers defining the shape of the array.

Method add_new_sparse_ndarray()

Add a new SOMA SparseNdArray to this collection. (lifecycle: experimental)

Usage

SOMACollectionBase$add_new_sparse_ndarray(key, type, shape)

Arguments

key

The key to be added.

type

an Arrow type defining the type of each element in the array.

shape

a vector of integers defining the shape of the array.

Method clone()

The objects of this class are cloneable with this method.

Usage

SOMACollectionBase$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.