SOMA NDArray Base Class

Adds NDArray-specific functionality to the SOMAArrayBase class. (lifecycle: experimental)

Super classes

tiledbsoma::TileDBObject -> tiledbsoma::TileDBArray -> tiledbsoma::SOMAArrayBase -> SOMANDArrayBase

Methods

Public methods

SOMANDArrayBase$create()

SOMANDArrayBase$clone()

Inherited methods

tiledbsoma::TileDBObject$class()

tiledbsoma::TileDBObject$exists()

tiledbsoma::TileDBObject$initialize()

tiledbsoma::TileDBObject$is_open()

tiledbsoma::TileDBObject$mode()

tiledbsoma::TileDBArray$attributes()

tiledbsoma::TileDBArray$attrnames()

tiledbsoma::TileDBArray$close()

tiledbsoma::TileDBArray$colnames()

tiledbsoma::TileDBArray$dimensions()

tiledbsoma::TileDBArray$dimnames()

tiledbsoma::TileDBArray$fragment_count()

tiledbsoma::TileDBArray$get_metadata()

tiledbsoma::TileDBArray$index_column_names()

tiledbsoma::TileDBArray$ndim()

tiledbsoma::TileDBArray$non_empty_domain()

tiledbsoma::TileDBArray$open()

tiledbsoma::TileDBArray$print()

tiledbsoma::TileDBArray$schema()

tiledbsoma::TileDBArray$set_metadata()

tiledbsoma::TileDBArray$shape()

tiledbsoma::TileDBArray$tiledb_array()

tiledbsoma::TileDBArray$tiledb_schema()

tiledbsoma::TileDBArray$used_shape()

Method create()

Create a SOMA NDArray named with the URI. (lifecycle: experimental)

Usage

SOMANDArrayBase$create(
  type,
  shape,
  platform_config = NULL,
  internal_use_only = NULL
)

Arguments

type

an Arrow type defining the type of each element in the array.

shape

a vector of integers defining the shape of the array.

platform_config

A platform configuration object

internal_use_only

Character value to signal this is a 'permitted' call, as create() is considered internal and should not be called directly.

Method clone()

The objects of this class are cloneable with this method.

Usage

SOMANDArrayBase$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.