SOMASparseNDArray

SOMASparseNDArray is a sparse, N-dimensional array with offset (zero-based) integer indexing on each dimension. The SOMASparseNDArray has a user-defined schema, which includes:

type - a primitive type, expressed as an Arrow type (e.g., int64, float32, etc)

shape - the shape of the array, i.e., number and length of each dimension

All dimensions must have a positive, non-zero length.

Note - on TileDB this is an sparse array with N int64 dimensions of domain [0, maxInt64), and a single attribute.

Duplicate writes

As duplicate index values are not allowed, index values already present in the object are overwritten and new index values are added. (lifecycle: experimental)

Super classes

tiledbsoma::TileDBObject -> tiledbsoma::TileDBArray -> tiledbsoma::SOMAArrayBase -> SOMASparseNDArray

Methods

Public methods

SOMASparseNDArray$create()

SOMASparseNDArray$read_arrow_table()

SOMASparseNDArray$read_sparse_matrix_zero_based()

SOMASparseNDArray$write()

SOMASparseNDArray$nnz()

SOMASparseNDArray$clone()

Inherited methods

tiledbsoma::TileDBObject$class()

tiledbsoma::TileDBObject$exists()

tiledbsoma::TileDBObject$initialize()

tiledbsoma::TileDBObject$is_open()

tiledbsoma::TileDBObject$mode()

tiledbsoma::TileDBArray$attributes()

tiledbsoma::TileDBArray$attrnames()

tiledbsoma::TileDBArray$close()

tiledbsoma::TileDBArray$colnames()

tiledbsoma::TileDBArray$dimensions()

tiledbsoma::TileDBArray$dimnames()

tiledbsoma::TileDBArray$fragment_count()

tiledbsoma::TileDBArray$get_metadata()

tiledbsoma::TileDBArray$index_column_names()

tiledbsoma::TileDBArray$ndim()

tiledbsoma::TileDBArray$open()

tiledbsoma::TileDBArray$print()

tiledbsoma::TileDBArray$reset_query()

tiledbsoma::TileDBArray$schema()

tiledbsoma::TileDBArray$set_metadata()

tiledbsoma::TileDBArray$set_query()

tiledbsoma::TileDBArray$shape()

tiledbsoma::TileDBArray$tiledb_array()

tiledbsoma::TileDBArray$tiledb_schema()

tiledbsoma::SOMAArrayBase$read_complete()

tiledbsoma::SOMAArrayBase$read_next()

Method create()

Create a SOMASparseNDArray named with the URI. (lifecycle: experimental)

Usage

SOMASparseNDArray$create(
  type,
  shape,
  platform_config = NULL,
  internal_use_only = NULL
)

Arguments

type

an Arrow type defining the type of each element in the array.

shape

a vector of integers defining the shape of the array.

platform_config

A platform configuration object

internal_use_only

Character value to signal this is a 'permitted' call, as create() is considered internal and should not be called directly.

Method read_arrow_table()

Read as an 'arrow::Table' (lifecycle: experimental)

Usage

SOMASparseNDArray$read_arrow_table(
  coords = NULL,
  result_order = "auto",
  iterated = FALSE,
  log_level = "warn"
)

Arguments

coords

Optional list of integer vectors, one for each dimension, with a length equal to the number of values to read. If NULL, all values are read. List elements can be named when specifying a subset of dimensions.

result_order

Optional order of read results. This can be one of either "ROW_MAJOR, "COL_MAJOR", or "auto"` (default).

result_order

Optional order of read results. This can be one of either "ROW_MAJOR, "COL_MAJOR", or "auto"` (default).

iterated

Option boolean indicated whether data is read in call (when FALSE, the default value) or in several iterated steps.

log_level

Optional logging level with default value of "warn".

Returns

An arrow::Table.

Method read_sparse_matrix_zero_based()

Read as a sparse matrix (lifecycle: experimental)

Usage

SOMASparseNDArray$read_sparse_matrix_zero_based(
  coords = NULL,
  result_order = "auto",
  repr = c("C", "T", "R"),
  iterated = FALSE,
  log_level = "warn"
)

Arguments

coords

Optional list of integer vectors, one for each dimension, with a length equal to the number of values to read. If NULL, all values are read. List elements can be named when specifying a subset of dimensions.

result_order

Optional order of read results. This can be one of either "ROW_MAJOR, "COL_MAJOR", or "auto"` (default).

result_order

Optional order of read results. This can be one of either "ROW_MAJOR, "COL_MAJOR", or "auto"` (default).

repr

Optional one-character code for sparse matrix representation type

iterated

Option boolean indicated whether data is read in call (when FALSE, the default value) or in several iterated steps.

log_level

Optional logging level with default value of "warn".

Returns

A matrix-like object accessed using zero-based indexes. It supports only basic access operations with zero-based indexes as well as dim(), nrow(), and ncol(). Use as.one.based() to get a fully-featured sparse matrix object supporting more advanced operations (with one-based indexing).

Method write()

Write matrix-like data to the array. (lifecycle: experimental)

Usage

SOMASparseNDArray$write(values)

Arguments

values

Any matrix-like object coercible to a TsparseMatrix. Character dimension names are ignored because SOMANDArray's use integer indexing.

Method nnz()

Retrieve number of non-zero elements (lifecycle: experimental)

Usage

SOMASparseNDArray$nnz()

Returns

A scalar with the number of non-zero elements

Method clone()

The objects of this class are cloneable with this method.

Usage

SOMASparseNDArray$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.