SOMA Read Iterator over Arrow Table

TableReadIter is a class that allows for iteration over a reads on SOMASparseNDArray and SOMADataFrame. Iteration chunks are retrieved as arrow::Table

Super class

tiledbsoma::ReadIter -> TableReadIter

Methods

Public methods

TableReadIter$concat()

TableReadIter$clone()

Inherited methods

tiledbsoma::ReadIter$initialize()

tiledbsoma::ReadIter$read_complete()

tiledbsoma::ReadIter$read_next()

Method concat()

Concatenate remainder of iterator.

Usage

TableReadIter$concat()

Returns

arrow::Table

Method clone()

The objects of this class are cloneable with this method.

Usage

TableReadIter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.