TileDB Array Base Class

Base class for representing an individual TileDB array. (lifecycle: experimental)

Super class

tiledbsoma::TileDBObject -> TileDBArray

Active bindings

object

Access the underlying TileB object directly (either a tiledb::tiledb_array or tiledb::tiledb_group).

Methods

Public methods

TileDBArray$open()

TileDBArray$close()

TileDBArray$print()

TileDBArray$tiledb_array()

TileDBArray$get_metadata()

TileDBArray$set_metadata()

TileDBArray$schema()

TileDBArray$tiledb_schema()

TileDBArray$dimensions()

TileDBArray$shape()

TileDBArray$used_shape()

TileDBArray$non_empty_domain()

TileDBArray$ndim()

TileDBArray$attributes()

TileDBArray$dimnames()

TileDBArray$attrnames()

TileDBArray$colnames()

TileDBArray$index_column_names()

TileDBArray$fragment_count()

TileDBArray$clone()

Inherited methods

tiledbsoma::TileDBObject$class()

tiledbsoma::TileDBObject$exists()

tiledbsoma::TileDBObject$initialize()

tiledbsoma::TileDBObject$is_open()

tiledbsoma::TileDBObject$mode()

Method open()

Open the SOMA object for read or write.

Usage

TileDBArray$open(mode = c("READ", "WRITE"), internal_use_only = NULL)

Arguments

mode

Mode to open in; defaults to READ.

internal_use_only

Character value to signal this is a 'permitted' call, as open() is considered internal and should not be called directly.

Returns

The object, invisibly

Method close()

Close the SOMA object.

Usage

TileDBArray$close()

Returns

The object, invisibly

Method print()

Print summary of the array. (lifecycle: experimental)

Usage

TileDBArray$print()

Method tiledb_array()

Return a TileDBArray object (lifecycle: experimental)

Usage

TileDBArray$tiledb_array(...)

Arguments

...

Optional arguments to pass to tiledb::tiledb_array()

Returns

A tiledb::tiledb_array object.

Method get_metadata()

Retrieve metadata from the TileDB array. (lifecycle: experimental)

Usage

TileDBArray$get_metadata(key = NULL)

Arguments

key

The name of the metadata attribute to retrieve.

Returns

A list of metadata values.

Method set_metadata()

Add list of metadata to the specified TileDB array. (lifecycle: experimental)

Usage

TileDBArray$set_metadata(metadata)

Arguments

metadata

Named list of metadata to add.

Returns

NULL

Method schema()

Retrieve the array schema as an Arrow schema (lifecycle: experimental)

Usage

TileDBArray$schema()

Returns

A arrow::schema object

Method tiledb_schema()

Retrieve the array schema as TileDB schema (lifecycle: experimental)

Usage

TileDBArray$tiledb_schema()

Returns

A tiledb::tiledb_array_schema object

Method dimensions()

Retrieve the array dimensions (lifecycle: experimental)

Usage

TileDBArray$dimensions()

Returns

A named list of tiledb::tiledb_dim objects

Method shape()

Retrieve the shape, i.e. the capacity of each dimension. This will not necessarily match the bounds of occupied cells within the array. Rather, it is the bounds outside of which no data may be written. (lifecycle: experimental)

Usage

TileDBArray$shape()

Returns

A named vector of dimension length (and the same type as the dimension)

Method used_shape()

Retrieve the range of indexes for a dimension that were explicitly written.

Usage

TileDBArray$used_shape(simplify = FALSE, index1 = FALSE)

Arguments

simplify

Return a vector of bit64::integer64s containing only the upper bounds.

index1

Return the used shape with 1-based indices (0-based indices are returned by default)

Returns

A list containing the lower and upper bounds for the used shape. If simplify = TRUE, returns a vector of only the upper bounds.

Method non_empty_domain()

Retrieve the non-empty domain for each dimension. This method calls tiledb::tiledb_array_get_non_empty_domain_from_name for each dimension in the array.

Usage

TileDBArray$non_empty_domain(index1 = FALSE)

Arguments

index1

Return the non-empty domain with 1-based indices.

Returns

A vector of bit64::integer64s with one entry for each dimension.

Method ndim()

Retrieve number of dimensions (lifecycle: experimental)

Usage

TileDBArray$ndim()

Returns

A scalar with the number of dimensions

Method attributes()

Retrieve the array attributes (lifecycle: experimental)

Usage

TileDBArray$attributes()

Returns

A list of tiledb::tiledb_attr objects

Method dimnames()

Retrieve dimension names (lifecycle: experimental)

Usage

TileDBArray$dimnames()

Returns

A character vector with the array's dimension names

Method attrnames()

Retrieve attribute names (lifecycle: experimental)

Usage

TileDBArray$attrnames()

Returns

A character vector with the array's attribute names

Method colnames()

Retrieve the names of all columns, including dimensions and attributes (lifecycle: experimental)

Usage

TileDBArray$colnames()

Returns

A character vector with the array's column names

Method index_column_names()

Retrieve names of index (dimension) columns (lifecycle: experimental)

Usage

TileDBArray$index_column_names()

Returns

A character vector with the array index (dimension) names

Method fragment_count()

Get number of fragments in the array (lifecycle: experimental)

Usage

TileDBArray$fragment_count()

Method clone()

The objects of this class are cloneable with this method.

Usage

TileDBArray$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.