TileDB Group Base Class

Base class for interacting with TileDB groups (lifecycle: experimental)

Value

A TileDBArray or TileDBGroup.

Super class

tiledbsoma::TileDBObject -> TileDBGroup

Methods

Public methods

TileDBGroup$print()

TileDBGroup$create()

TileDBGroup$open()

TileDBGroup$close()

TileDBGroup$set()

TileDBGroup$get()

TileDBGroup$remove()

TileDBGroup$length()

TileDBGroup$names()

TileDBGroup$to_list()

TileDBGroup$to_data_frame()

TileDBGroup$get_metadata()

TileDBGroup$set_metadata()

TileDBGroup$clone()

Inherited methods

tiledbsoma::TileDBObject$class()

tiledbsoma::TileDBObject$exists()

tiledbsoma::TileDBObject$initialize()

tiledbsoma::TileDBObject$is_open()

tiledbsoma::TileDBObject$mode()

Method print()

Print summary of the group. (lifecycle: experimental)

Usage

TileDBGroup$print()

Method create()

Creates the data structure on disk/S3/cloud. (lifecycle: experimental)

Usage

TileDBGroup$create(internal_use_only = NULL)

Arguments

internal_use_only

Character value to signal this is a 'permitted' call, as create() is considered internal and should not be called directly.

Method open()

Open the SOMA object for read or write.

Usage

TileDBGroup$open(mode = c("READ", "WRITE"), internal_use_only = NULL)

Arguments

mode

Mode to open in; defaults to READ.

internal_use_only

Character value to signal this is a 'permitted' call, as open() is considered internal and should not be called directly.

Returns

The object, invisibly

Method close()

Close the SOMA object.

Usage

TileDBGroup$close()

Returns

The object, invisibly

Method set()

Add new member to the group. (lifecycle: experimental)

Usage

TileDBGroup$set(object, name = NULL, relative = NULL)

Arguments

object

A TileDBArray or TileDBGroup object to add.

name

Name to use for the member. By default the base name of the object's URI is used.

relative

An optional logical value indicating whether the new object's URI is relative to the group's URI. If NULL (the default), the object's URI is assumed to be relative unless it is a tiledb:// URI.

Method get()

Retrieve a group member by name. If the member isn't already open, it is opened in the same mode as the parent. (lifecycle: experimental)

Usage

TileDBGroup$get(name)

Arguments

name

The name of the member.

mode

Mode to open in

Method remove()

Remove member. (lifecycle: experimental)

Usage

TileDBGroup$remove(name)

Arguments

name

Name of the member to remove.

Method length()

Length in the number of members. (lifecycle: experimental)

Usage

TileDBGroup$length()

Returns

Scalar integer

Method names()

Retrieve the names of members. (lifecycle: experimental)

Usage

TileDBGroup$names()

Returns

A character vector of member names.

Method to_list()

Retrieve a list of members. (lifecycle: experimental)

Usage

TileDBGroup$to_list()

Method to_data_frame()

Retrieve a data.frame of members. (lifecycle: experimental)

Usage

TileDBGroup$to_data_frame()

Method get_metadata()

Retrieve metadata. (lifecycle: experimental)

Usage

TileDBGroup$get_metadata(key = NULL)

Arguments

key

The name of the metadata attribute to retrieve. is not NULL.

Returns

A list of metadata values.

Method set_metadata()

Add list of metadata. (lifecycle: experimental)

Usage

TileDBGroup$set_metadata(metadata)

Arguments

metadata

Named list of metadata to add.

Returns

NULL

Method clone()

The objects of this class are cloneable with this method.

Usage

TileDBGroup$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.