Iterator-Style Access to SOMA Array via SOMAArray

The sr_* functions provide low-level access to an instance of the SOMAArray class so that iterative access over parts of a (large) array is possible.

sr_setup

instantiates and by default also submits a query

sr_complete

checks if more data is available

sr_next

returns the next chunk

Usage

sr_setup(
  uri,
  config,
  colnames = NULL,
  qc = NULL,
  dim_points = NULL,
  dim_ranges = NULL,
  batch_size = "auto",
  result_order = "auto",
  timestamp_end = NULL,
  loglevel = "auto"
)

sr_complete(sr)

sr_next(sr)

Arguments

uri

Character value with URI path to a SOMA data set

config

Named chracter vector with ‘key’ and ‘value’ pairs used as TileDB config parameters.

colnames

Optional vector of character value with the name of the columns to retrieve

qc

Optional external Pointer object to TileDB Query Condition, defaults to ‘NULL’ i.e. no query condition

dim_points

Optional named list with vector of data points to select on the given dimension(s). Each dimension can be one entry in the list.

dim_ranges

Optional named list with two-column matrix where each row select a range for the given dimension. Each dimension can be one entry in the list.

batch_size

Optional argument for size of data batches, defaults to ‘auto’

result_order

Optional argument for query result order, defaults to ‘auto’

timestamp_end

Optional POSIXct (i.e. Datetime) type for end of interval for which data is considered.

loglevel

Character value with the desired logging level, defaults to ‘auto’ which lets prior setting prevail, any other value is set as new logging level.

sr

An external pointer to a TileDB SOMAArray object

Value

sr_setup returns an external pointer to a SOMAArray. sr_complete

returns a boolean, and sr_next returns an Arrow array helper object.

Examples

if (FALSE) {
uri <- extract_dataset("soma-dataframe-pbmc3k-processed-obs")
ctx <- tiledb::tiledb_ctx()
sr <- sr_setup(uri, config=as.character(tiledb::config(ctx)))
rl <- data.frame()
while (!sr_complete(sr)) {
  dat <- sr_next(sr)
  rb <- arrow::RecordBatch$import_from_c(dat$array_data, dat$schema)
  rl <- rbind(rl, as.data.frame(rb))
}
summary(rl)
}