Write R Objects to SOMA

Various helpers to write R objects to SOMA

Usage

# S3 method for class 'DataFrame'
write_soma(
  x,
  uri,
  soma_parent,
  df_index = NULL,
  index_column_names = "soma_joinid",
  ...,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

# S3 method for class 'Hits'
write_soma(
  x,
  uri,
  soma_parent,
  sparse = TRUE,
  type = NULL,
  transpose = FALSE,
  ...,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

# S3 method for class 'character'
write_soma(
  x,
  uri,
  soma_parent,
  ...,
  key = NULL,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

# S3 method for class 'data.frame'
write_soma(
  x,
  uri,
  soma_parent,
  df_index = NULL,
  index_column_names = "soma_joinid",
  ...,
  key = NULL,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

# S3 method for class 'matrix'
write_soma(
  x,
  uri,
  soma_parent,
  sparse = TRUE,
  type = NULL,
  transpose = FALSE,
  ...,
  key = NULL,
  ingest_mode = "write",
  shape = NULL,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

# S3 method for class 'Matrix'
write_soma(
  x,
  uri,
  soma_parent,
  sparse = TRUE,
  type = NULL,
  transpose = FALSE,
  ...,
  key = NULL,
  ingest_mode = "write",
  shape = NULL,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

# S3 method for class 'TsparseMatrix'
write_soma(
  x,
  uri,
  soma_parent,
  type = NULL,
  transpose = FALSE,
  ...,
  key = NULL,
  ingest_mode = "write",
  shape = NULL,
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

Arguments

x

An object

uri

URI for resulting SOMA object

soma_parent

The parent collection (eg. a SOMACollection, SOMAExperiment, or SOMAMeasurement)

df_index

The name of the column in x with the index (row names); by default, will automatically add the row names of x to an attribute named “index” to the resulting SOMADataFrame

index_column_names

Names of columns in x to index in the resulting SOMA object

...

Arguments passed to other methods

ingest_mode

Ingestion mode when creating the SOMA; choose from:

“write”: create a new SOMA and error if it already exists

“resume”: attempt to create a new SOMA; if it already exists, simply open it for writing

platform_config

Optional platform configuration

tiledbsoma_ctx

Optional SOMATileDBContext

relative

[Internal use only] Is uri relative or aboslute

sparse

Create a sparse or dense array from x

type

Arrow type for encoding x (eg. arrow::int32()); by default, attempts to determine arrow type with arrow::infer_type()

transpose

Transpose x before writing

key

Optionally register the resulting SOMADataFrame in soma_parent as key; pass NULL to prevent registration to handle manually

shape

A vector of two positive integers giving the on-disk shape of the array; defaults to dim(x)

Value

The resulting SOMA array or data frame, returned opened for write

Writing Character Arrays

Characters are written out as SOMADataFrames with one attribute: “values”; additionally one bit of array-level metadata is added:

“soma_uns_outgest_hint ” with a value of “array_1d ”

Writing Data Frames

Data frames are written out as SOMADataFrames. The following transformations are applied to x:

row names are added to a column in x entitled “index”, “_index”, or a random name if either option is already present in x

a column “soma_joinid” will be automatically added going from [0, nrow(x) - 1] encoded as 64-bit integers

The array type for each column will be determined by arrow::infer_type(); if any column contains a non-atomic type (excluding factors, complexes,and raws), the code will error out

Writing Dense Matrices

Dense matrices are written as two-dimensional dense arrays. The overall shape of the array is determined by dim(x) and the type of the array is determined by type or arrow::infer_type(x)

Writing Sparse Matrices

Sparse matrices are written out as two-dimensional TileDB sparse arrays in COO format:

the row indices (“i”) are written out as “soma_dim_0”

the column indices (“j”) are written out as “soma_dim_1”

the non-zero values (“x”) are written out as “soma_data”

The array type is determined by type, or arrow::infer_type(slot(x, "x"))