Convert a Seurat Sub-Object to a SOMA Object, returned opened for write

Various helpers to write Seurat sub-objects to SOMA objects.

Usage

# S3 method for class 'Assay'
write_soma(
  x,
  uri = NULL,
  soma_parent,
  ...,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

# S3 method for class 'DimReduc'
write_soma(
  x,
  uri = NULL,
  soma_parent,
  fidx = NULL,
  nfeatures = NULL,
  ...,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

# S3 method for class 'Graph'
write_soma(
  x,
  uri,
  soma_parent,
  ...,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

# S3 method for class 'SeuratCommand'
write_soma(
  x,
  uri = NULL,
  soma_parent,
  ...,
  ingest_mode = "write",
  platform_config = NULL,
  tiledbsoma_ctx = NULL,
  relative = TRUE
)

Arguments

x

A Seurat sub-object (eg. an Assay, a DimReduc, or a Graph)

uri

URI for resulting SOMA object

soma_parent

The parent collection; for the DimReduc and Graph methods, this must be a measurement for the assay x was generated from

...

Arguments passed to other methods

ingest_mode

Ingestion mode when creating the SOMA; choose from:

“write”: create a new SOMA and error if it already exists

“resume”: attempt to create a new SOMA; if it already exists, simply open it for writing

platform_config

Optional platform configuration

tiledbsoma_ctx

Optional SOMATileDBContext

relative

[Internal use only] Is uri relative or aboslute

fidx

An integer vector describing the location of features in SeuratObject::Loadings(x) with relation to soma_parent (eg. match(rownames(Loadings(x)), rownames(assay)))

nfeatures

The number of features present in soma_parent (eg. nrow(assay))

Value

Assay method: a SOMAMeasurement with the data from x, returned opened for write

DimReduc and Graph methods: invisibly returns soma_parent, opened for write, with the values of x added to it

Writing Assays

Seurat Assay objects are written out as individual measurements:

the “data” matrix is written out a sparse matrix called “data” within the “X” group

the “counts” matrix, if not empty, is written out a sparse matrix called “counts” within the “X” group

the “scale.data” matrix, if not empty, is written out a sparse matrix called “scale_data” within the “X” group

feature-level meta data is written out as a data frame called “var”

Expression matrices are transposed (cells as rows) prior to writing. All other slots, including results from extended assays (eg. SCTAssay, ChromatinAssay) are lost

Writing DimReducs

Seurat DimReduc objects are written out to the “obsm” and “varm” groups of a measurement:

cell embeddings are written out as a sparse matrix in the “obsm” group

feature loadings, if not empty, are written out as a sparse matrix in the “varm” groups; loadings are padded with NAs to include all features

Dimensional reduction names are translated to AnnData-style names (eg. “pca” becomes X_pca for embeddings and “PCs” for loadings). All other slots, including projected feature loadings and jackstraw information, are lost

Writing Graphs

Seurat Graph objects are written out as sparse matrices to the “obsp” group of a measurement

Writing SeuratCommands

Seurat command logs are written out as data frames to the “seurat_commands” group of a collection